Refactor p_commands.py
