+++ docs/source/api_ref/hli/ports/ports.rst
@@ -4,6 +4,7 @@
 Traffic Generation APIs
 ------------------------
 
+
 Port APIs Summary Table
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
@@ -11,6 +12,7 @@ Port APIs Summary Table
     :header: "Id", "Action", "API", "Description"
     :widths: auto
     :file: ports_tg.csv
+
 
 Port API Examples
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
