+++ docs/source/api_ref/lli/index.rst
@@ -9,17 +9,20 @@
     c_commands
     m_commands
     p_commands
+    p_macsec_commands
+    p_lldp_commands
+    p_ue_commands
     ps_commands
     pt_commands
     pr_commands
-    pp_commands
     pc_commands
     pd_commands
     pm_commands
-    pf_commands
     pl_commands
-    px_commands
+    pf_commands
+    pp_commands
     pl1_commands
+    px_commands
     pe_commands
     pec_commands
     ped_commands
@@ -49,17 +52,20 @@
     submodules/c_commands
     submodules/m_commands
     submodules/p_commands
+    submodules/p_macsec_commands
+    submodules/p_lldp_commands
+    submodules/p_ue_commands
     submodules/ps_commands
     submodules/pt_commands
     submodules/pr_commands
-    submodules/pp_commands
     submodules/pc_commands
     submodules/pd_commands
     submodules/pm_commands
-    submodules/pf_commands
     submodules/pl_commands
-    submodules/px_commands
+    submodules/pf_commands
+    submodules/pp_commands
     submodules/pl1_commands
+    submodules/px_commands
     submodules/pe_commands
     submodules/pec_commands
     submodules/ped_commands
