+++ docs/source/api_ref/lli/submodules/p_commands.rst
@@ -6,8 +6,8 @@
 This module contains the **L23 port classes** that deal with basic information about, and configuration of L23 test ports. The L23 port command names all have the form ``P_<xxx>`` and require a module index id and a port index id. In general, port commands cannot be changed while traffic is on. Additionally, every stream must be disabled before changing parameters that affect the bandwidth of the port.
 
 
-Port Commands
------------------
+Port Commands - Traffic Generation
+-----------------------------------
 
 .. automodule:: xoa_driver.internals.commands.p_commands
     :members:
