+++ docs/source/api_ref/lli/submodules/p_macsec_commands.rst
@@ -1,0 +1,13 @@
+``p_macsec_commands`` module
+=================================
+
+This module contains the **L23 port classes (MACsec)** that deal with basic information about, and configuration of L23 test ports. The L23 port command names all have the form ``P_MACSEC_<xxx>`` and require a module index id and a port index id. 
+
+
+Port Commands - MACsec
+----------------------
+
+.. automodule:: xoa_driver.internals.commands.p_macsec_commands
+    :members:
+    :no-undoc-members:
+    :exclude-members: __init__
