+++ xoa_driver/internals/commands/p_commands.py
@@ -1,4 +1,4 @@
-"""Port Commands"""
+"""Port Commands - Traffic Generation"""
 from __future__ import annotations
 from dataclasses import dataclass
 import ipaddress
@@ -63,24 +63,6 @@ from .enums import (
     TrafficError,
     TrafficEngine,
     ReconciliationSublayerSupport,
-    OnOff,
-    MACSecSCIMode,
-    MACSecCipherSuite,
-    # MACSecVLANMode,
-    MACSecRekeyMode,
-    MACSecEncryptionMode,
-    MACSecPNMode,
-    LLDPClearTarget,
-    LLDPOpMode,
-    UecCtlosClearDirection,
-    UecLinkOptionLlr,
-    UecLlrBehavior,
-    UecLlrInitEchoMode,
-    UecLlrEchoCheck,
-    UecLlrTxErrType,
-    UecLlrTxErrPattern,
-    UecLlrTxFsmState,
-    UecLlrRxFsmState,
 )
 
 
@@ -4755,2054 +4737,6 @@ class P_FAULTCNT:
 
         return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
 
-
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_CREATE:
-    """
-    Create a TX Secure Channel (SC) on the port.
-    """
-
-    code: typing.ClassVar[int] = 505
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-
-    class SetDataAttr(RequestBodyStruct):
-        pass
-
-    def set(self) -> Token[None]:
-        """Create a TX Secure Channel (SC) on the port.
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_INDICES:
-    """
-    Create multiple TX SCs or query the existing TX SCs on the port.
-    """
-
-    code: typing.ClassVar[int] = 506
-    pushed: typing.ClassVar[bool] = True
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        txsc_indices: typing.List[int] = field(XmpSequence(types_chunk=[XmpInt()]))
-        """list of integers, the sub-indices of TX SCs on the port."""
-
-    class SetDataAttr(RequestBodyStruct):
-        txsc_indices: typing.List[int] = field(XmpSequence(types_chunk=[XmpInt()]))
-        """list of integers, the sub-indices of TX SCs on the port."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the full list of which TX SCs are defined for a port.
-
-        :return: the sub-indices of TX SCs on the port
-        :rtype: P_MACSEC_TXSC_INDICES.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-    def set(self, txsc_indices: typing.List[int]) -> Token[None]:
-        """Creates a new empty TX SC for each value that is not already in use, and deletes each TX SC that is not mentioned in the list.
-
-        :param txsc_indices: the sub-indices of TX SCs on the port
-        :type txsc_indices: typing.List[int]
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, txsc_indices=txsc_indices))
-
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_DELETE:
-    """
-    Delete a TX Secure Channel (SC) on the port.
-    """
-
-    code: typing.ClassVar[int] = 530
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-
-    class SetDataAttr(RequestBodyStruct):
-        pass
-
-    def set(self) -> Token[None]:
-        """Delete a TX Secure Channel (SC) on the port.
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_CONF_OFFSET:
-    """
-    The confidentiality offset of the port’s TX SC.
-    """
-
-    code: typing.ClassVar[int] = 510
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        offset: int = field(XmpInt())
-        """integer, the TX Secure Channel (SC) offset. Allowed values are 0, 30, and 50"""
-
-    class SetDataAttr(RequestBodyStruct):
-        offset: int = field(XmpInt())
-        """integer, the TX Secure Channel (SC) offset. Allowed values are 0, 30, and 50"""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the TX Secure Channel (SC) offset on the port.
-
-        :return: the TX Secure Channel (SC) offset on the port
-        :rtype: P_MACSEC_TXSC_CONF_OFFSET.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-
-    def set(self, offset: int) -> Token[None]:
-        """Set the TX Secure Channel (SC) offset on the port.
-
-        :param offset: the TX Secure Channel (SC) offset
-        :type offset: int
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index], offset=offset))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_DESCR:
-    """
-    The description of the port’s TX SC.
-    """
-
-    code: typing.ClassVar[int] = 507
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        description: str = field(XmpStr())
-        """string, the description of the TX Secure Channel (SC)."""
-
-    class SetDataAttr(RequestBodyStruct):
-        description: str = field(XmpStr())
-        """string, the description of the TX Secure Channel (SC)."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the description of the TX Secure Channel (SC) on the port.
-
-        :return: the description of the TX Secure Channel (SC) on the port
-        :rtype: P_MACSEC_TXSC_DESCR.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-
-    def set(self, description: str) -> Token[None]:
-        """Set the description of the TX Secure Channel (SC) on the port.
-
-        :param description: the description of the TX Secure Channel (SC)
-        :type description: str
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index], description=description))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_SCI_MODE:
-    """
-    The mode of the port’s TX SCI in MACsec.
-
-    """
-
-    code: typing.ClassVar[int] = 513
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        mode: MACSecSCIMode = field(XmpByte())
-        """coded byte, the mode of the port’s TX SCI in MACsec."""
-
-    class SetDataAttr(RequestBodyStruct):
-        mode: MACSecSCIMode = field(XmpByte())
-        """coded byte, the mode of the port’s TX SCI in MACsec."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the mode of the port’s TX SCI in MACsec.
-
-        :return: the mode of the port’s TX SCI in MACsec.
-        :rtype: P_MACSEC_TXSC_SCI_MODE.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-
-    def set(self, mode: MACSecSCIMode) -> Token[None]:
-        """Set the mode of the port’s TX SCI in MACsec.
-
-        :param mode: the mode of the port’s TX SCI in MACsec.
-        :type mode: MACSecSCIMode
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index], mode=mode))
-
-    set_end_station = functools.partialmethod(set, MACSecSCIMode.END_STATION)
-    """Set SCI Mode to END STATION.
-    """
-
-    set_with_sci = functools.partialmethod(set, MACSecSCIMode.WITH_SCI)
-    """Set SCI Mode to WITH SCI.
-    """
-
-    # set_no_sci = functools.partialmethod(set, MACSecSCIMode.NO_SCI)
-    # """Set SCI Mode to NO SCI.
-    # """
-
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_SCI:
-    """
-    The SCI of the port’s TX SC.
-    """
-
-    code: typing.ClassVar[int] = 508
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        sci: Hex = field(XmpHex(size=8))
-        """hex 8 bytes, the SCI of the port’s TX SC."""
-
-    class SetDataAttr(RequestBodyStruct):
-        sci: Hex = field(XmpHex(size=8))
-        """hex 8 bytes, the SCI of the port’s TX SC."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the SCI of the port’s TX SC.
-
-        :return: the SCI of the port’s TX SC.
-        :rtype: P_MACSEC_TXSC_SCI.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-
-    def set(self, sci: Hex) -> Token[None]:
-        """Set the SCI of the port’s TX SC.
-
-        :param sci: The SCI of the port’s TX SC.
-        :type sci: int
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index], sci=sci))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_CIPHERSUITE:
-    """
-    The cipher suite of the port’s TX SC.
-    """
-
-    code: typing.ClassVar[int] = 509
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        cipher_suite: MACSecCipherSuite = field(XmpByte())
-        """coded byte, the cipher suite of the port’s TX SC."""
-
-    class SetDataAttr(RequestBodyStruct):
-        cipher_suite: MACSecCipherSuite = field(XmpByte())
-        """coded byte, the cipher suite of the port’s TX SC."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the cipher suite of the port’s TX SC.
-
-        :return: the cipher suite of the port’s TX SC.
-        :rtype: P_MACSEC_TXSC_CIPHERSUITE.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-
-    def set(self, cipher_suite: MACSecCipherSuite) -> Token[None]:
-        """Set the cipher suite of the port’s TX SC.
-
-        :param cipher_suite: the cipher suite of the port’s TX SC.
-        :type cipher_suite: MACSecCipherSuite
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index], cipher_suite=cipher_suite))
-
-    set_gcm_aes_128 = functools.partialmethod(set, MACSecCipherSuite.GCM_AES_128)
-    """Set cipher suite to GCM_AES_128.
-    """
-
-    set_gcm_aes_256 = functools.partialmethod(set, MACSecCipherSuite.GCM_AES_256)
-    """Set cipher suite to GCM_AES_256.
-    """
-
-    set_gcm_aes_xpn_128 = functools.partialmethod(set, MACSecCipherSuite.GCM_AES_XPN_128)
-    """Set cipher suite to GCM_AES_XPN_128.
-    """
-
-    set_gcm_aes_xpn_256 = functools.partialmethod(set, MACSecCipherSuite.GCM_AES_XPN_256)
-    """Set cipher suite to GCM_AES_XPN_256.
-    """
-
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_STARTING_PN:
-    """
-    The starting PN number of the port’s TX SC uses.
-    """
-
-    code: typing.ClassVar[int] = 514
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        starting_pn: int = field(XmpLong())
-        """integer, the starting PN number. Default to 1, maximum 2^64. Allowed to be 0."""
-
-        mode: MACSecPNMode = field(XmpByte())
-        """byte, defining how to continue the TX PN after the start-traffic. Default to CONTINUOUS."""
-
-    class SetDataAttr(RequestBodyStruct):
-        starting_pn: int = field(XmpLong())
-        """integer, the starting PN number. Default to 1, maximum 2^64. Allowed to be 0."""
-
-        mode: MACSecPNMode = field(XmpByte())
-        """byte, defining how to continue the TX PN after the start-traffic. Default to CONTINUOUS."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the starting PN number. Default to 1, maximum 2^64. Allowed to be 0.
-
-        :return: the starting PN number. Default to 1, maximum 2^64.
-        :rtype: P_MACSEC_TXSC_STARTING_PN.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-
-    def set(self, starting_pn: int, mode: MACSecPNMode) -> Token[None]:
-        """Set the starting PN number. Default to 1, maximum 2^64. Allowed to be 0.
-
-        :param starting_pn: the starting PN number. Default to 1, maximum 2^64.
-        :type starting_pn: int
-        :param mode: defining how to continue the TX PN after the start-traffic.
-        :type mode: MACSecPNMode
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index], starting_pn=starting_pn, mode=mode))
-    
-
-# @register_command
-# @dataclass
-# class P_MACSEC_TXSC_VLAN_MODE:
-#     """
-#     The VLAN mode of the port’s TX SC.
-    
-#         * VLAN encrypted: The original MACsec header format encoded the 802.1Q tag as part of the encrypted payload, thus hiding it from the public Ethernet transport.
-        
-#         * VLAN in clear text (WAN MACsec): With 802.1Q tag in the clear, the 802.1Q tag is encoded outside the 802.1AE encryption header, exposing the tag to the private and public Ethernet transport.
-
-#     """
-
-#     code: typing.ClassVar[int] = 511
-#     pushed: typing.ClassVar[bool] = False
-
-#     _connection: 'interfaces.IConnection'
-#     _module: int
-#     _port: int
-#     _txsc_index: int
-
-#     class GetDataAttr(ResponseBodyStruct):
-#         mode: MACSecVLANMode = field(XmpByte())
-#         """integer, the VLAN mode. Default to ENCRYPTED."""
-
-#     class SetDataAttr(RequestBodyStruct):
-#         mode: MACSecVLANMode = field(XmpByte())
-#         """integer, the VLAN mode. Default to ENCRYPTEDC."""
-
-#     def get(self) -> Token[GetDataAttr]:
-#         """Get the VLAN mode.
-
-#         :return: the VLAN mode.
-#         :rtype: P_MACSEC_TXSC_VLAN_MODE.GetDataAttr
-#         """
-#         return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-
-#     def set(self, mode: MACSecVLANMode) -> Token[None]:
-#         """Set the VLAN mode.
-
-#         :param mode: the VLAN mode.
-#         :type mode: MACSecVLANMode
-#         """
-#         return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index], mode=mode))
-    
-#     set_encrypted = functools.partialmethod(set, MACSecVLANMode.ENCRYPTED)
-#     """Set VLAN mode to ENCRYPTED.
-#     """
-
-#     set_clear_text = functools.partialmethod(set, MACSecVLANMode.CLEAR_TEXT)
-#     """Set VLAN mode to CLEAR_TEXT.
-#     """
-
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_REKEY_MODE:
-    """
-    The rekey mode of the port’s TX SC defines when to switch to the next SAK.
-    """
-
-    code: typing.ClassVar[int] = 515
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        mode: MACSecRekeyMode = field(XmpByte())
-        """byte, the rekey mode of the port’s TX SC"""
-
-        value: int = field(XmpInt())
-        """integer, defines the packet number that triggers the rekey. This value will be ignored if the mode is set to PN_EXHAUSTION"""
-
-    class SetDataAttr(RequestBodyStruct):
-        mode: MACSecRekeyMode = field(XmpByte())
-        """byte, the rekey mode of the port’s TX SC"""
-
-        value: int = field(XmpInt())
-        """integer, defines the packet number that triggers the rekey. This value will be ignored if the mode is set to PN_EXHAUSTION"""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the rekey mode of the port’s TX SC
-
-        :return: the rekey mode of the port’s TX SC
-        :rtype: P_MACSEC_TXSC_REKEY_MODE.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-
-    def set(self, mode: MACSecRekeyMode, value: int) -> Token[None]:
-        """Set the rekey mode.
-
-        :param mode: the rekey mode.
-        :type mode: MACSecRekeyMode
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index], mode=mode, value=value))
-
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_ENCRYPT:
-    """
-    The encryption mode of the port’s TX SC.
-    """
-
-    code: typing.ClassVar[int] = 512
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        mode: MACSecEncryptionMode = field(XmpByte())
-        """byte, the encryption mode of the port’s TX SC"""
-
-    class SetDataAttr(RequestBodyStruct):
-        mode: MACSecEncryptionMode = field(XmpByte())
-        """byte, the encryption mode of the port’s TX SC"""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the encryption mode of the port’s TX SC
-
-        :return: the encryption mode of the port’s TX SC
-        :rtype: P_MACSEC_TXSC_ENCRYPT.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-
-    def set(self, mode: MACSecEncryptionMode) -> Token[None]:
-        """Set the encryption mode.
-
-        :param mode: the encryption mode.
-        :type mode: MACSecEncryptionMode
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index], mode=mode))
-    
-    set_encrypt_integrity = functools.partialmethod(set, MACSecEncryptionMode.ENCRYPT_INTEGRITY)
-    """Set encryption mode to encryption and integrity.
-    """
-
-    set_integrity_only = functools.partialmethod(set, MACSecEncryptionMode.INTEGRITY_ONLY)
-    """Set encryption mode to integrity only.
-    """
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_SAK_VALUE:
-    """
-    Configure the value of a SAK key on the port’s TX SC.
-    """
-
-    code: typing.ClassVar[int] = 534
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-    _sak_key_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        sak_key_value: Hex = field(XmpHex())
-        """integer, the SAK key. Default to all-zero. Allowed to be empty."""
-
-    class SetDataAttr(RequestBodyStruct):
-        sak_key_value: Hex = field(XmpHex())
-        """integer, the SAK key. Default to all-zero. Allowed to be empty."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the SAK key.
-
-        :return: the the SAK key.
-        :rtype: P_MACSEC_TXSC_SAK_VALUE.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._txsc_index, self._sak_key_index]))
-
-    def set(self, sak_key_value: Hex) -> Token[None]:
-        """Set the SAK key.
-
-        :param sak_key_value: the SAK key.
-        :type sak_key_value: Hex
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index, self._sak_key_index], sak_key_value=sak_key_value))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_XPN_SSCI:
-    """
-    The XPN Short SCI of the port's TX SC when XPN cipher suite is in use.
-    """
-
-    code: typing.ClassVar[int] = 540
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        xpn_ssci: Hex = field(XmpHex(size=4))
-        """hex 4 bytes, The XPN Short SCI of the port's TX SC when XPN cipher suite is in use"""
-
-    class SetDataAttr(RequestBodyStruct):
-        xpn_ssci: Hex = field(XmpHex(size=4))
-        """hex 4 bytes, The XPN Short SCI of the port's TX SC when XPN cipher suite is in use"""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the XPN Short SCI of the port's TX SC when XPN cipher suite is in use
-
-        :return: the SCI of the port’s TX SC.
-        :rtype: P_MACSEC_TXSC_XPN_SSCI.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-
-    def set(self, xpn_ssci: Hex) -> Token[None]:
-        """Set the XPN Short SCI of the port's TX SC when XPN cipher suite is in use
-
-        :param xpn_ssci: the XPN Short SCI of the port's TX SC when XPN cipher suite is in use
-        :type xpn_ssci: int
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index], xpn_ssci=xpn_ssci))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_XPN_SALT:
-    """
-    The XPN salt of the port's TX SC when XPN cipher suite is in use.
-    """
-
-    code: typing.ClassVar[int] = 541
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        xpn_salt: Hex = field(XmpHex(size=12))
-        """hex 12 bytes, XPN salt of the port's TX SC when XPN cipher suite is in use."""
-
-    class SetDataAttr(RequestBodyStruct):
-        xpn_salt: Hex = field(XmpHex(size=12))
-        """hex 12 bytes, XPN salt of the port's TX SC when XPN cipher suite is in use."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get XPN salt of the port's TX SC when XPN cipher suite is in use.
-
-        :return: the SCI of the port’s TX SC.
-        :rtype: P_MACSEC_TXSC_XPN_SALT.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-
-    def set(self, xpn_salt: Hex) -> Token[None]:
-        """Set XPN salt of the port's TX SC when XPN cipher suite is in use.
-
-        :param xpn_salt: XPN salt of the port's TX SC when XPN cipher suite is in use.
-        :type xpn_salt: int
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index], xpn_salt=xpn_salt))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_CREATE:
-    """
-    Create a RX Secure Channel (SC) on the port.
-    """
-
-    code: typing.ClassVar[int] = 518
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _rxsc_index: int
-
-    class SetDataAttr(RequestBodyStruct):
-        pass
-
-    def set(self) -> Token[None]:
-        """Create a RX Secure Channel (SC) on the port.
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._rxsc_index]))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_INDICES:
-    """
-    Create multiple RX SCs or query the existing RX SCs on the port.
-    """
-
-    code: typing.ClassVar[int] = 519
-    pushed: typing.ClassVar[bool] = True
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        rxsc_indices: typing.List[int] = field(XmpSequence(types_chunk=[XmpInt()]))
-        """list of integers, the sub-indices of RX SCs on the port."""
-
-    class SetDataAttr(RequestBodyStruct):
-        rxsc_indices: typing.List[int] = field(XmpSequence(types_chunk=[XmpInt()]))
-        """list of integers, the sub-indices of RX SCs on the port."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the full list of which RX SCs are defined for a port.
-
-        :return: the sub-indices of RX SCs on the port
-        :rtype: P_MACSEC_RXSC_INDICES.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-    def set(self, rxsc_indices: typing.List[int]) -> Token[None]:
-        """Creates a new empty RX SC for each value that is not already in use, and deletes each RX SC that is not mentioned in the list.
-
-        :param rxsc_indices: the sub-indices of RX SCs on the port
-        :type rxsc_indices: typing.List[int]
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, rxsc_indices=rxsc_indices))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_DELETE:
-    """
-    Delete a RX Secure Channel (SC) on the port.
-    """
-
-    code: typing.ClassVar[int] = 531
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _rxsc_index: int
-
-    class SetDataAttr(RequestBodyStruct):
-        pass
-
-    def set(self) -> Token[None]:
-        """Delete a RX Secure Channel (SC) on the port.
-        """
-        
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._rxsc_index]))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_DESCR:
-    """
-    The description of the port’s RX SC.
-    """
-
-    code: typing.ClassVar[int] = 520
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _rxsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        description: str = field(XmpStr())
-        """string, the description of the RX Secure Channel (SC)."""
-
-    class SetDataAttr(RequestBodyStruct):
-        description: str = field(XmpStr())
-        """string, the description of the RX Secure Channel (SC)."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the description of the RX Secure Channel (SC) on the port.
-
-        :return: the description of the RX Secure Channel (SC) on the port
-        :rtype: P_MACSEC_RXSC_DESCR.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._rxsc_index]))
-
-    def set(self, description: str) -> Token[None]:
-        """Set the description of the RX Secure Channel (SC) on the port.
-
-        :param description: the description of the RX Secure Channel (SC)
-        :type description: str
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._rxsc_index], description=description))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_SCI:
-    """
-    The SCI of the port’s RX SC.
-    """
-
-    code: typing.ClassVar[int] = 523
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _rxsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        sci: Hex = field(XmpHex(size=8))
-        """hex 8 bytes, the SCI of the port’s RX SC."""
-
-    class SetDataAttr(RequestBodyStruct):
-        sci: Hex = field(XmpHex(size=8))
-        """hex 8 bytes, the SCI of the port’s RX SC."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the SCI of the port’s RX SC.
-
-        :return: the SCI of the port’s RX SC.
-        :rtype: P_MACSEC_RXSC_SCI.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._rxsc_index]))
-
-    def set(self, sci: Hex) -> Token[None]:
-        """Set the SCI of the port’s RX SC.
-
-        :param sci: The SCI of the port’s RX SC.
-        :type sci: int
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._rxsc_index], sci=sci))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_CONF_OFFSET:
-    """
-    The confidentiality offset of the port’s RX SC.
-    """
-
-    code: typing.ClassVar[int] = 522
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _rxsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        offset: int = field(XmpInt())
-        """integer, the RX Secure Channel (SC) offset. Allowed values are 0, 30, and 50."""
-
-    class SetDataAttr(RequestBodyStruct):
-        offset: int = field(XmpInt())
-        """integer, the RX Secure Channel (SC) offset. Allowed values are 0, 30, and 50"""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the RX Secure Channel (SC) offset on the port.
-
-        :return: the RX Secure Channel (SC) offset on the port
-        :rtype: P_MACSEC_RXSC_CONF_OFFSET.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._rxsc_index]))
-
-    def set(self, offset: int) -> Token[None]:
-        """Set the RX Secure Channel (SC) offset on the port.
-
-        :param offset: the RX Secure Channel (SC) offset
-        :type offset: int
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._rxsc_index], offset=offset))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_CIPHERSUITE:
-    """
-    The cipher suite of the port’s RX SC.
-    """
-
-    code: typing.ClassVar[int] = 521
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _rxsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        cipher_suite: MACSecCipherSuite = field(XmpByte())
-        """coded byte, the cipher suite of the port’s RX SC."""
-
-    class SetDataAttr(RequestBodyStruct):
-        cipher_suite: MACSecCipherSuite = field(XmpByte())
-        """coded byte, the cipher suite of the port’s RX SC."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the cipher suite of the port’s RX SC.
-
-        :return: the cipher suite of the port’s RX SC.
-        :rtype: P_MACSEC_RXSC_CIPHERSUITE.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._rxsc_index]))
-
-    def set(self, cipher_suite: MACSecCipherSuite) -> Token[None]:
-        """Set the cipher suite of the port’s RX SC.
-
-        :param cipher_suite: the cipher suite of the port’s RX SC.
-        :type cipher_suite: MACSecCipherSuite
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._rxsc_index], cipher_suite=cipher_suite))
-
-    set_gcm_aes_128 = functools.partialmethod(set, MACSecCipherSuite.GCM_AES_128)
-    """Set cipher suite to GCM_AES_128.
-    """
-
-    set_gcm_aes_256 = functools.partialmethod(set, MACSecCipherSuite.GCM_AES_256)
-    """Set cipher suite to GCM_AES_256.
-    """
-
-    set_gcm_aes_xpn_128 = functools.partialmethod(set, MACSecCipherSuite.GCM_AES_XPN_128)
-    """Set cipher suite to GCM_AES_XPN_128.
-    """
-
-    set_gcm_aes_xpn_256 = functools.partialmethod(set, MACSecCipherSuite.GCM_AES_XPN_256)
-    """Set cipher suite to GCM_AES_XPN_256.
-    """
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_LOWEST_PN:
-    """
-    The lowest PN number of the port’s RX SC expects to receive.
-    """
-
-    code: typing.ClassVar[int] = 511
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _rxsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        lowest_pn: Hex = field(XmpHex(size=8))
-        """hex, The lowest PN number of the port’s RX SC expects to receive. Default to 1, maximum 2^64. Allowed to be 0."""
-
-    class SetDataAttr(RequestBodyStruct):
-        lowest_pn: Hex = field(XmpHex(size=8))
-        """hex, The lowest PN number of the port’s RX SC expects to receive. Default to 1, maximum 2^64. Allowed to be 0."""
-
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the lowest PN number of the port’s RX SC expects to receive
-
-        :return: the lowest PN number of the port’s RX SC expects to receive
-        :rtype: P_MACSEC_RXSC_LOWEST_PN.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._rxsc_index]))
-
-    def set(self, lowest_pn: int) -> Token[None]:
-        """Set the lowest PN number of the port’s RX SC expects to receive
-
-        :param lowest_pn: The lowest PN number of the port’s RX SC expects to receive. Default to 1, maximum 2^64.
-        :type lowest_pn: int
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._rxsc_index], lowest_pn=lowest_pn))
-
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_TPLDID:
-    """
-    Associate a TPLD ID with the RX SC.
-    """
-
-    code: typing.ClassVar[int] = 535
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _rxsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        tpld_id: int = field(XmpInt())
-        """integer, the TPLD ID to associate with the RX SC."""
-
-    class SetDataAttr(RequestBodyStruct):
-        tpld_id: int = field(XmpInt())
-        """integer, the TPLD ID to associate with the RX SC."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the TPLD ID to associate with the RX SC.
-
-        :return: the TPLD ID to associate with the RX SC.
-        :rtype: P_MACSEC_RXSC_TPLDID.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._rxsc_index]))
-
-    def set(self, tpld_id: int) -> Token[None]:
-        """Set the TPLD ID to associate with the RX SC.
-
-        :param tpld_id: the TPLD ID to associate with the RX SC.
-        :type tpld_id: int
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._rxsc_index], tpld_id=tpld_id))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_SAK_VALUE:
-    """
-    Configure the value of a SAK key on the port’s RX SC.
-    """
-
-    code: typing.ClassVar[int] = 542
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _rxsc_index: int
-    _sak_key_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        sak_key_value: Hex = field(XmpHex())
-        """integer, the SAK key. Default to all-zero. Allowed to be empty."""
-
-    class SetDataAttr(RequestBodyStruct):
-        sak_key_value: Hex = field(XmpHex())
-        """integer, the SAK key. Default to all-zero. Allowed to be empty."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the SAK key.
-
-        :return: the the SAK key.
-        :rtype: P_MACSEC_RXSC_SAK_VALUE.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._rxsc_index, self._sak_key_index]))
-
-    def set(self, sak_key_value: Hex) -> Token[None]:
-        """Set the SAK key.
-
-        :param sak_key_value: the SAK key.
-        :type sak_key_value: Hex
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._rxsc_index, self._sak_key_index], sak_key_value=sak_key_value))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_XPN_SSCI:
-    """
-    The XPN Short SCI of the port's RX SC when XPN cipher suite is in use.
-    """
-
-    code: typing.ClassVar[int] = 544
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _rxsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        xpn_ssci: Hex = field(XmpHex(size=4))
-        """hex 4 bytes, The XPN Short SCI of the port's RX SC when XPN cipher suite is in use"""
-
-    class SetDataAttr(RequestBodyStruct):
-        xpn_ssci: Hex = field(XmpHex(size=4))
-        """hex 4 bytes, The XPN Short SCI of the port's RX SC when XPN cipher suite is in use"""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the XPN Short SCI of the port's RX SC when XPN cipher suite is in use
-
-        :return: the SCI of the port’s RX SC.
-        :rtype: P_MACSEC_RXSC_XPN_SSCI.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._rxsc_index]))
-
-    def set(self, xpn_ssci: Hex) -> Token[None]:
-        """Set the XPN Short SCI of the port's RX SC when XPN cipher suite is in use
-
-        :param xpn_ssci: the XPN Short SCI of the port's RX SC when XPN cipher suite is in use
-        :type xpn_ssci: int
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._rxsc_index], xpn_ssci=xpn_ssci))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_XPN_SALT:
-    """
-    The XPN salt of the port's RX SC when XPN cipher suite is in use.
-    """
-
-    code: typing.ClassVar[int] = 546
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _rxsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        xpn_salt: Hex = field(XmpHex(size=12))
-        """hex 12 bytes, XPN salt of the port's RX SC when XPN cipher suite is in use."""
-
-    class SetDataAttr(RequestBodyStruct):
-        xpn_salt: Hex = field(XmpHex(size=12))
-        """hex 12 bytes, XPN salt of the port's RX SC when XPN cipher suite is in use."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get XPN salt of the port's RX SC when XPN cipher suite is in use.
-
-        :return: the SCI of the port’s RX SC.
-        :rtype: P_MACSEC_TXSC_XPN_SALT.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._rxsc_index]))
-
-    def set(self, xpn_salt: Hex) -> Token[None]:
-        """Set XPN salt of the port's RX SC when XPN cipher suite is in use.
-
-        :param xpn_salt: XPN salt of the port's RX SC when XPN cipher suite is in use.
-        :type xpn_salt: int
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._rxsc_index], xpn_salt=xpn_salt))
-
-
-@register_command
-@dataclass
-class P_MACSEC_TX_STATS:
-    """
-    Port-level MACsec TX counters
-    """
-
-    code: typing.ClassVar[int] = 517
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        bits_sec: int = field(XmpLong())
-        """long integer, the number of MACsec L2 bits transmitted of the previous second."""
-        bytes_sec: int = field(XmpLong())
-        """long integer, the number of MACsec L2 bytes transmitted of the previous second."""
-        frames_sec: int = field(XmpLong())
-        """long integer, the number of MACsec frames transmitted of the previous second."""
-        total_bits: int = field(XmpLong())
-        """long integer, the number of MACsec L2 bits transmitted since last cleared."""
-        total_bytes: int = field(XmpLong())
-        """long integer, the number of MACsec L2 bytes transmitted since last cleared."""
-        total_frames: int = field(XmpLong())
-        """long integer, the number of MACsec frames transmitted since last cleared."""
-        total_protected_only_bits: int = field(XmpLong())
-        """long integer, the number of protected-only (non-encrypted) bits transmitted by the port since cleared."""
-        total_protected_only_bytes: int = field(XmpLong())
-        """long integer, the number of protected-only (non-encrypted) bytes transmitted by the port since cleared."""
-        total_encrypted_bits: int = field(XmpLong())
-        """long integer, the number of encrypted bits transmitted by the port since cleared, excluding the bytes in the Confidentiality Offset."""
-        total_encrypted_bytes: int = field(XmpLong())
-        """long integer, the number of encrypted bytes transmitted by the port since cleared, excluding the bytes in the Confidentiality Offset."""
-
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get port-level MACsec TX counters
-
-        :return: port-level MACsec TX counters
-        :rtype: P_MACSEC_TX_STATS.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_STATS:
-    """
-    SC/stream-level MACsec TX counters.
-    """
-
-    code: typing.ClassVar[int] = 528
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        bits_sec: int = field(XmpLong())
-        """long integer, the number of MACsec L2 bits transmitted of the previous second."""
-        bytes_sec: int = field(XmpLong())
-        """long integer, the number of MACsec L2 bytes transmitted of the previous second."""
-        frames_sec: int = field(XmpLong())
-        """long integer, the number of MACsec frames transmitted of the previous second."""
-        total_bits: int = field(XmpLong())
-        """long integer, the number of MACsec L2 bits transmitted since last cleared."""
-        total_bytes: int = field(XmpLong())
-        """long integer, the number of MACsec L2 bytes transmitted since last cleared."""
-        total_frames: int = field(XmpLong())
-        """long integer, the number of MACsec frames transmitted since last cleared."""
-        total_protected_only_bits: int = field(XmpLong())
-        """long integer, the number of protected-only (non-encrypted) bits transmitted since cleared."""
-        total_protected_only_bytes: int = field(XmpLong())
-        """long integer, the number of protected-only (non-encrypted) bytes transmitted since cleared."""
-        total_encrypted_bits: int = field(XmpLong())
-        """long integer, the number of encrypted bits transmitted since cleared, excluding the bytes in the Confidentiality Offset."""
-        total_encrypted_bytes: int = field(XmpLong())
-        """long integer, the number of encrypted bytes transmitted since cleared, excluding the bytes in the Confidentiality Offset."""
-
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get SC/stream-level MACsec TX counters.
-
-        :return: SC/stream-level MACsec TX counters.
-        :rtype: P_MACSEC_TXSC_STATS.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_TX_CLEAR:
-    """
-    Clear the MACsec TX counters of the port.
-    """
-
-    code: typing.ClassVar[int] = 516
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class SetDataAttr(RequestBodyStruct):
-        pass
-
-    def set(self) -> Token[None]:
-        """Clear the MACsec TX counters of the port.
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port))
-    
-
-
-@register_command
-@dataclass
-class P_MACSEC_RX_STATS:
-    """
-    Port-level MACsec RX counters
-    """
-
-    code: typing.ClassVar[int] = 525
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        bits_sec: int = field(XmpLong())
-        """long integer, number of MACsec L2 bits received of the previous second."""
-        bytes_sec: int = field(XmpLong())
-        """long integer, number of MACsec L2 bytes received of the previous second."""
-        frames_sec: int = field(XmpLong())
-        """long integer, number of MACsec frames received of the previous second."""
-        total_bits: int = field(XmpLong())
-        """long integer, number of MACsec L2 bits received since last cleared."""
-        total_bytes: int = field(XmpLong())
-        """long integer, number of MACsec L2 bytes received since last cleared."""
-        total_frames: int = field(XmpLong())
-        """long integer, number of MACsec frames received since last cleared."""
-        total_ok_frames: int = field(XmpLong())
-        """long integer, the number of good MACsec frames received since cleared."""
-        total_delayed_frames: int = field(XmpLong())
-        """long integer, the number of frames with the PN lower than the minmum expected since cleared."""
-        total_icv_check_failed_frames: int = field(XmpLong())
-        """long integer, the number of frames with ICV check failed recevied since cleared."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get port-level MACsec RX counters
-
-        :return: port-level MACsec RX counters
-        :rtype: P_MACSEC_RX_STATS.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_STATS:
-    """
-    SC/stream-level MACsec RX counters
-    """
-
-    code: typing.ClassVar[int] = 529
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _rxsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        bits_sec: int = field(XmpLong())
-        """long integer, number of MACsec L2 bits received of the previous second."""
-        bytes_sec: int = field(XmpLong())
-        """long integer, number of MACsec L2 bytes received of the previous second."""
-        frames_sec: int = field(XmpLong())
-        """long integer, number of MACsec frames received of the previous second."""
-        total_bits: int = field(XmpLong())
-        """long integer, number of MACsec L2 bits received since last cleared."""
-        total_bytes: int = field(XmpLong())
-        """long integer, number of MACsec L2 bytes received since last cleared."""
-        total_frames: int = field(XmpLong())
-        """long integer, number of MACsec frames received since last cleared."""
-        total_ok_frames: int = field(XmpLong())
-        """long integer, the number of good MACsec frames received since cleared."""
-        total_delayed_frames: int = field(XmpLong())
-        """long integer, the number of frames with the PN lower than the minmum expected since cleared."""
-        total_icv_check_failed_frames: int = field(XmpLong())
-        """long integer, the number of frames with ICV check failed recevied since cleared."""
-
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get port-level MACsec RX counters
-
-        :return: port-level MACsec RX counters
-        :rtype: P_MACSEC_RX_STATS.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._rxsc_index]))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_RX_CLEAR:
-    """
-    Clear the MACsec RX counters of the port.
-    """
-
-    code: typing.ClassVar[int] = 524
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class SetDataAttr(RequestBodyStruct):
-        pass
-
-    def set(self) -> Token[None]:
-        """Clear the MACsec RX counters of the port.
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_RX_ENABLE:
-    """
-    This will enable/disable the MACSec functionality on the RX side. With it ON, the RX port will try to decode the received packets. If it is OFF, the port will not try to decode any received packets.
-    """
-
-    code: typing.ClassVar[int] = 545
-    pushed: typing.ClassVar[bool] = True
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        on_off: OnOff = field(XmpByte())
-        """coded byte, enable or disable MACsec on the RX port."""
-
-    class SetDataAttr(RequestBodyStruct):
-        on_off: OnOff = field(XmpByte())
-        """coded byte, enable or disable MACsec on the RX port."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the RX port MACSec state.
-
-        :return: the RX port MACSec stat
-        :rtype: PS_MACSEC_ENABLE.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-    def set(self, on_off: OnOff) -> Token[None]:
-        """Set the RX port MACSec stat.
-
-        :param on_off: the RX port MACSec stat
-        :type on_off: OnOff
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, on_off=on_off))
-
-    set_off = functools.partialmethod(set, OnOff.OFF)
-    """Disable the RX port MACSec.
-    """
-
-    set_on = functools.partialmethod(set, OnOff.ON)
-    """Enable the RX port MACSec.
-    """
-
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_AN:
-    """
-    RX SC's next AN
-    """
-
-    code: typing.ClassVar[int] = 550
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _rxsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        next_an: int = field(XmpInt())
-        """integer, the next AN"""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get RX SC's next AN
-
-        :return: RX SC's next AN
-        :rtype: P_MACSEC_RXSC_AN.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._rxsc_index]))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_NEXT_PN:
-    """
-    The next PN number of the port’s RX SC expects to receive.
-    """
-
-    code: typing.ClassVar[int] = 548
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _rxsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        next_pn: Hex = field(XmpHex(size=8))
-        """hex, The next PN number of the port’s RX SC expects to receive. Default to 1, maximum 2^64. Allowed to be 0."""
-
-    class SetDataAttr(RequestBodyStruct):
-        next_pn: Hex = field(XmpHex(size=8))
-        """hex, The next PN number of the port’s RX SC expects to receive. Default to 1, maximum 2^64. Allowed to be 0."""
-
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the next PN number of the port’s RX SC expects to receive
-
-        :return: the next PN number of the port’s RX SC expects to receive
-        :rtype: P_MACSEC_RXSC_NEXT_PN.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._rxsc_index]))
-
-    def set(self, value: int) -> Token[None]:
-        """Set the next PN number of the port’s RX SC expects to receive
-
-        :param value: The next PN number of the port’s RX SC expects to receive. Default to 1, maximum 2^64.
-        :type value: int
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._rxsc_index], value=value))
-    
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_NEXT_PN:
-    """
-    The next PN number of the port’s TX SC expects to receive.
-    """
-
-    code: typing.ClassVar[int] = 547
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        next_pn: Hex = field(XmpHex(size=8))
-        """hex, The next PN number of the port’s TX SC expects to receive. Default to 1, maximum 2^64. Allowed to be 0."""
-    class SetDataAttr(RequestBodyStruct):
-        next_pn: Hex = field(XmpHex(size=8))
-        """hex, The next PN number of the port’s TX SC expects to receive. Default to 1, maximum 2^64. Allowed to be 0."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the next PN number of the port’s TX SC expects to receive
-
-        NextPN is a monotonically incrementing 32 or 64 bit counter that is never zero. 
-
-        :return: the next PN number of the port’s TX SC expects to receive
-        :rtype: P_MACSEC_TXSC_NEXT_PN.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-
-    def set(self, next_pn: int) -> Token[None]:
-        """Set the next PN number of the port’s TX SC expects to receive
-
-        NextPN is a monotonically incrementing 32 or 64 bit counter that is never zero. 
-
-        :param next_pn: The next PN number of the port’s TX SC expects to receive. Default to 1, maximum 2^64.
-        :type next_pn: int
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index], next_pn=next_pn))
-    
-
-
-@register_command
-@dataclass
-class P_MACSEC_RXSC_PN:
-    """
-    RX SC's next PN
-    """
-
-    code: typing.ClassVar[int] = 551
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _rxsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        pn: Hex = field(XmpHex(size=8))
-        """hex, 8-byte long, the latest recovered PN for the Rx SC"""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get RX SC's next PN
-        
-        :return: RX SC's next PN
-        :rtype: P_MACSEC_RXSC_PN.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._rxsc_index]))
-
-
-@register_command
-@dataclass
-class P_MACSEC_TXSC_NEXT_AN:
-    """
-    The next AN number of the port’s TX SC expects to receive.
-    """
-
-    code: typing.ClassVar[int] = 549
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _txsc_index: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        next_an: int = field(XmpInt())
-        """integer, The next AN number of the port’s TX SC expects to receive.  Allowed to be 0."""
-
-    class SetDataAttr(RequestBodyStruct):
-        next_an: int = field(XmpInt())
-        """integer, The next AN number of the port’s TX SC expects to receive.  Allowed to be 0."""
-    
-    def get(self) -> Token[GetDataAttr]:
-        """Get the next AN number of the port’s TX SC expects to receive
-
-        :return: the next AN number of the port’s TX SC expects to receive
-        :rtype: P_MACSEC_TXSC_NEXT_AN.GetDataAttr
-        """
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._txsc_index]))
-
-    def set(self, next_an: int) -> Token[None]:
-        """Set the next AN number of the port’s TX SC expects to receive
-
-        :param next_an: The next AN number of the port’s TX SC expects to receive. 
-        :type next_an: int
-        """
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._txsc_index], next_an=next_an))
-    
-
-
-@register_command
-@dataclass
-class P_LLDP_CLEAR:
-    """
-    Clear LLDP information for the specified target.
-    """
-
-    code: typing.ClassVar[int] = 1009
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-
-    class SetDataAttr(RequestBodyStruct):
-        target: LLDPClearTarget = field(XmpByte())
-        """LLDP clear target."""
-
-
-    def set(self, target: LLDPClearTarget) -> Token[None]:
-        """Set the LLDP clear target.
-
-        :param target: the LLDP clear target
-        :type target: LLDPClearTarget
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, target=target))
-    
-    clear_none = functools.partialmethod(set, LLDPClearTarget.NONE)
-    """Not clear LLDP stats or neighbors
-    """
-
-    clear_neighbor = functools.partialmethod(set, LLDPClearTarget.NEIGHBOR)
-    """Clear LLDP neighbor information, but not LLDP statistics.
-    """
-
-    clear_stats = functools.partialmethod(set, LLDPClearTarget.STATS)
-    """Clear LLDP statistics, but not LLDP neighbor information.
-    """
-
-    clear_all = functools.partialmethod(set, LLDPClearTarget.ALL)
-    """Clear both LLDP statistics and neighbor information.
-    """
-    
-
-@register_command
-@dataclass
-class P_LLDP_CONFIG:
-    """
-    LLDP agent configuration for the port, including reinit delay, transmission interval, and transmission hold multiplier.
-    """
-
-    code: typing.ClassVar[int] = 1006
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _lldp_agent_xindex: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        
-        reinit_delay: int = field(XmpInt())
-        """Reinitialization delay in seconds. When a port is disabled, LLDP is disabled or the switch is rebooted, a LLDP shutdown frame is transmitted to the neighboring units, signaling that the LLDP information is not valid anymore. Tx Reinit controls the amount of seconds between the shutdown frame and a new LLDP initialization. 
-        
-        Default is 2, range is 1 to 10.
-        """
-
-        tx_interval: int = field(XmpInt())
-        """Transmission interval in seconds. This variable defines the time interval in seconds between transmissions during normal transmission periods. 
-        
-        Default is 30, range is 1 to 3600.
-        """
-
-        tx_hold_multiplier: int = field(XmpInt())
-        """Transmission hold multiplier. The number of times the transmission interval is multiplied to determine the hold time for the LLDP packet. It is also used to determine the value of TTL that is carried in LLDP frames transmitted the port.
-        
-        Default is 4, range is 1 to 100.
-        """
-
-        tx_delay: int = field(XmpInt())
-        """Minimum time between LLDP frames in seconds. If some configuration is changed (for example, the IP address) a new LLDP frame is transmitted, but the time between the LLDP frames will always be at least the value of ``tx_delay`` seconds. ``tx_delay`` cannot be larger than 1/4 of the ``tx_interval`` value. 
-        
-        Valid values are restricted to 1 - 8192 seconds.
-        """
-
-    class SetDataAttr(RequestBodyStruct):
-        reinit_delay: int = field(XmpInt())
-        """Reinitialization delay in seconds. When a port is disabled, LLDP is disabled or the switch is rebooted, a LLDP shutdown frame is transmitted to the neighboring units, signaling that the LLDP information is not valid anymore. Tx Reinit controls the amount of seconds between the shutdown frame and a new LLDP initialization. 
-        
-        Default is 2, range is 1 to 10.
-        """
-
-        tx_interval: int = field(XmpInt())
-        """Transmission interval in seconds. This variable defines the time interval in seconds between transmissions during normal transmission periods. 
-        
-        Default is 30, range is 1 to 3600.
-        """
-
-        tx_hold_multiplier: int = field(XmpInt())
-        """Transmission hold multiplier. The number of times the transmission interval is multiplied to determine the hold time for the LLDP packet. It is also used to determine the value of TTL that is carried in LLDP frames transmitted the port.
-        
-        Default is 4, range is 1 to 100.
-        """
-
-        tx_delay: int = field(XmpInt())
-        """Minimum time between LLDP frames in seconds. If some configuration is changed (for example, the IP address) a new LLDP frame is transmitted, but the time between the LLDP frames will always be at least the value of ``tx_delay`` seconds. ``tx_delay`` cannot be larger than 1/4 of the ``tx_interval`` value. 
-        
-        Valid values are restricted to 1 - 8192 seconds.
-        """
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the LLDP configuration for the port.
-
-        :return: the LLDP configuration for the port
-        :rtype: P_LLDP_CONFIG.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._lldp_agent_xindex]))
-
-    def set(self, reinit_delay: int, tx_interval: int, tx_hold_multiplier: int, tx_delay: int) -> Token[None]:
-        """Set the LLDP configuration for the port.
-
-        :param reinit_delay: Reinitialization delay in seconds
-        :type reinit_delay: int
-        :param tx_interval: Transmission interval in seconds
-        :type tx_interval: int
-        :param tx_hold_multiplier: Transmission hold multiplier
-        :type tx_hold_multiplier: int
-        :param tx_delay: Minimum time between LLDP frames in seconds
-        :type tx_delay: int
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._lldp_agent_xindex], reinit_delay=reinit_delay, tx_interval=tx_interval, tx_hold_multiplier=tx_hold_multiplier, tx_delay=tx_delay))
-
-
-@register_command
-@dataclass
-class P_LLDP_CREATE:
-    """
-    Create a LLDP agent on the port.
-    """
-
-    code: typing.ClassVar[int] = 1000
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _lldp_agent_xindex: int
-
-    class SetDataAttr(RequestBodyStruct):
-        pass
-
-    def set(self) -> Token[None]:
-        """Create a LLDP agent on the port.
-
-        LLDP agent index, ranges from 0 to 7. (maximum 8 LLDP agents)
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._lldp_agent_xindex]))
-    
-
-
-@register_command
-@dataclass
-class P_LLDP_DATA:
-    """
-    Configures the LLDPDU (LLDP Data Unit) for a specified LLDP agent on the port.
-    """
-
-    code: typing.ClassVar[int] = 1004
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _lldp_agent_xindex: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        
-        data_unit: Hex = field(XmpHex())
-        """LLDPDU in hexadecimal string format."""
-
-    class SetDataAttr(RequestBodyStruct):
-        
-        data_unit: Hex = field(XmpHex())
-        """LLDPDU in hexadecimal string format."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the LLDPDU for the specified LLDP agent on the port.
-
-        :return: the LLDPDU for the specified LLDP agent on the port
-        :rtype: P_LLDP_DATA.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._lldp_agent_xindex]))
-
-    def set(self, data_unit: Hex) -> Token[None]:
-        """Set the LLDPDU for the specified LLDP agent on the port.
-
-        :param data_unit: the LLDPDU in hexadecimal string format
-        :type data_unit: Hex
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._lldp_agent_xindex], data_unit=data_unit))
-    
-
-@register_command
-@dataclass
-class P_LLDP_DELETE:
-    """
-    Delete a LLDP agent on the port.
-    """
-
-    code: typing.ClassVar[int] = 1002
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _lldp_agent_xindex: int
-
-    class SetDataAttr(RequestBodyStruct):
-        pass
-
-    def set(self) -> Token[None]:
-        """Delete a LLDP agent on the port.
-
-        LLDP agent index, ranges from 0 to 7. (maximum 8 LLDP agents)
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._lldp_agent_xindex]))
-    
-
-
-@register_command
-@dataclass
-class P_LLDP_HEADER:
-    """
-    Configures the LLDP header (DMAC, SMAC, EtherType) for a specified LLDP agent on the port.
-    """
-
-    code: typing.ClassVar[int] = 1005
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _lldp_agent_xindex: int
-
-
-    class GetDataAttr(ResponseBodyStruct):
-        
-        header: Hex = field(XmpHex())
-        """LLDP frame header in hexadecimal string format. The header includes DMAC, SMAC, and Ethertype.
-        
-        * The default DMAC is the MAC address of the test port. If the port MAC address is changed, the default DMAC will also be updated to match the new port MAC address.
-        * The default SMAC is Nearest Bridge MAC address, ``0x0180C200000E``, as specified in IEEE 802.3-2018.
-        * The default Ethertype is ``0x88CC`` for LLDP frames, as specified in IEEE 802.3-2018.
-        
-        """
-
-    class SetDataAttr(RequestBodyStruct):
-        
-        header: Hex = field(XmpHex())
-        """LLDP frame header in hexadecimal string format. The header includes DMAC, SMAC, and Ethertype.
-        
-        * The default DMAC is the MAC address of the test port. If the port MAC address is changed, the default DMAC will also be updated to match the new port MAC address.
-        * The default SMAC is Nearest Bridge MAC address, ``0x0180C200000E``, as specified in IEEE 802.3-2018.
-        * The default Ethertype is ``0x88CC`` for LLDP frames, as specified in IEEE 802.3-2018.
-        
-        """
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the LLDP header.
-
-        :return: the LLDP header 
-        :rtype: P_LLDP_HEADER.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._lldp_agent_xindex]))
-
-    def set(self, header: Hex) -> Token[None]:
-        """Set the description of the port.
-
-        :param header: the LLDP header in hexadecimal string format
-        :type header: Hex
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._lldp_agent_xindex], header=header))
-    
-
-@register_command
-@dataclass
-class P_LLDP_INDICES:
-    """
-    Create multiple LLDP agents or query the existing LLDP agents on the port.
-    """
-
-    code: typing.ClassVar[int] = 1001
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        
-        lldp_agent_indices: typing.List[int] = field(XmpSequence(types_chunk=[XmpInt()]))
-        """List of LLDP agent indices on the port. Each index ranges from 0 to 7, and a maximum of 8 LLDP agents can be created on the port. The integers do not have to be consecutive. If the list is empty, it indicates that there are no LLDP agents on the port."""
-
-    class SetDataAttr(RequestBodyStruct):
-        
-        lldp_agent_indices: typing.List[int] = field(XmpSequence(types_chunk=[XmpInt()]))
-        """List of LLDP agent indices on the port. Each index ranges from 0 to 7, and a maximum of 8 LLDP agents can be created on the port. The integers do not have to be consecutive. If the list is empty, it indicates that there are no LLDP agents on the port."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the indices of the LLDP agents on the port.
-
-        :return: the indices of the LLDP agents on the port
-        :rtype: P_LLDP_INDICES.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-    def set(self, lldp_agent_indices: typing.List[int]) -> Token[None]:
-        """Set the indices of the LLDP agents on the port.
-
-        :param lldp_agent_indices: the list of LLDP agent indices
-        :type lldp_agent_indices: typing.List[int]
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, lldp_agent_indices=lldp_agent_indices))
-
-
-@register_command
-@dataclass
-class P_LLDP_NEIGHBORS:
-    """
-    Displays LLDP neighbors discovered by the port.
-    """
-
-    code: typing.ClassVar[int] = 1007
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        
-        neighbors: dict = field(XmpJson(min_len=16))
-        """Information of LLDP neighbors."""
-
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the LLDP neighbors discovered by the port.
-
-        :return: the LLDP neighbors discovered by the port
-        :rtype: P_LLDP_NEIGHBORS.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-
-
-@register_command
-@dataclass
-class P_LLDP_OPMODE:
-    """
-    This configures the LLDP operational mode for the port.
-    """
-
-    code: typing.ClassVar[int] = 1003
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-    _lldp_agent_xindex: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        
-        op_mode: LLDPOpMode = field(XmpByte())
-        """LLDP operational mode for the port."""
-
-    class SetDataAttr(RequestBodyStruct):
-        
-        op_mode: LLDPOpMode = field(XmpByte())
-        """LLDP operational mode for the port."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the LLDP operational mode of the port.
-
-        :return: the LLDP operational mode of the port
-        :rtype: P_LLDP_OPMODE.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._lldp_agent_xindex]))
-
-    def set(self, op_mode: LLDPOpMode) -> Token[None]:
-        """Set the LLDP operational mode of the port.
-
-        :param op_mode: the LLDP operational mode of the port
-        :type op_mode: LLDPOpMode
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, indices=[self._lldp_agent_xindex], op_mode=op_mode))
-
-    set_disable = functools.partialmethod(set, LLDPOpMode.DISABLE)
-    """Disable LLDP on the port. The port will not transmit LLDP frames, and any received LLDP frames will be discarded."""
-
-    set_tx_only = functools.partialmethod(set, LLDPOpMode.TX_ONLY)
-    """Enable LLDP transmission only. The port will transmit LLDP frames, but any received LLDP frames will be discarded."""
-
-    set_rx_only = functools.partialmethod(set, LLDPOpMode.RX_ONLY)
-    """Enable LLDP reception only. The port will not transmit LLDP frames, but it will process any received LLDP frames and learn neighbors."""
-
-    set_tx_rx = functools.partialmethod(set, LLDPOpMode.TX_RX)
-    """Enable both LLDP transmission and reception on the port. The port will transmit LLDP frames, and it will also process any received LLDP frames and learn neighbors."""
-
-
-@register_command
-@dataclass
-class P_LLDP_STATS:
-    """
-    Displays LLDP interface statistics for the port (all agents on the port combined).
-    """
-
-    code: typing.ClassVar[int] = 1008
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        
-        tx_frame_count: int = field(XmpLong())
-        """Number of LLDP frames transmitted."""
-
-        rx_frame_count: int = field(XmpLong())
-        """Number of LLDP frames received."""
-
-        rx_frame_discard_count: int = field(XmpLong())
-        """Number of LLDP frames discarded. LLDP frames that contain detectable errors in the first three mandatory TLVs are discarded."""
-
-        rx_tlv_discarded: int = field(XmpLong())
-        """Number of TLVs discarded. Optional TLVs that contain detectable errors are discarded."""
-
-        rx_tlv_unrecognized: int = field(XmpLong())
-        """Number of unrecognized TLVs received. An unrecognized TLV is referred to as the TLV whose type value is in the range of reserved TLV types. An unrecognized TLV may be a basic management TLV from a later LLDP version."""
-
-        rx_aged_out_count: int = field(XmpLong())
-        """Number of aged out. This counter provides a count of the times that a neighbor’s information has been deleted due to TTL expiration."""
-
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the description of the port.
-
-        :return: the description of the port
-        :rtype: P_COMMENT.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-
 @register_command
 @dataclass
 class P_EDUN_RX_STATUS:
@@ -6872,1309 +4806,6 @@ class P_EDUN_RX_STATUS:
         """
 
         return Token(self._connection, build_get_request(self, module=self._module, port=self._port, indices=[self._serdes_xindex]))
-
-
-
-@register_command
-@dataclass
-class P_UE_CTLOS_CLEAR:
-    """
-    Clear UE CtlOS counters in the specified direction(s).
-    """
-
-    code: typing.ClassVar[int] = 1020
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-
-    class SetDataAttr(RequestBodyStruct):
-        direction: UecCtlosClearDirection = field(XmpByte())
-        """UE CtlOS clear direction."""
-
-
-    def set(self, direction: UecCtlosClearDirection) -> Token[None]:
-        """Set the UE CtlOS clear direction.
-
-        :param direction: the UE CtlOS clear direction
-        :type direction: UecCtlosClearDirection
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, direction=direction))
-    
-    clear_none = functools.partialmethod(set, UecCtlosClearDirection.NONE)
-    """Not clear UE CtlOS counters
-    """
-
-    clear_rx = functools.partialmethod(set, UecCtlosClearDirection.RX)
-    """Clear RX UE CtlOS counters
-    """
-
-    clear_tx = functools.partialmethod(set, UecCtlosClearDirection.TX)
-    """Clear TX UE CtlOS counters
-    """
-
-    clear_all = functools.partialmethod(set, UecCtlosClearDirection.ALL)
-    """Clear all UE CtlOS counters (RX and TX).
-    """
-    
-
-@register_command
-@dataclass
-class P_UE_CTLOS_RX_STATS:
-    """
-    CtlOS Rx statistics, including counters of all UE CtlOS message type values.
-    """
-
-    code: typing.ClassVar[int] = 1012
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        
-        llr_init_cnt: int = field(XmpLong())
-        """The number of LLR_INIT CtlOS received."""
-
-        llr_init_echo_cnt: int = field(XmpLong())
-        """The number of LLR_INIT_ECHO CtlOS received."""
-
-        llr_ack_cnt: int = field(XmpLong())
-        """The number of LLR_ACK CtlOS received."""
-
-        llr_nack_cnt: int = field(XmpLong())
-        """The number of LLR_NACK CtlOS received."""
-
-        llr_init_seq: Hex = field(XmpHex(size=3))
-        """The LLR_INIT sequence number. 3 bytes in hex format."""
-
-        llr_init_data: Hex = field(XmpHex(size=2))
-        """The LLR_INIT data. 2 bytes in hex format."""
-
-        llr_init_echo_seq: Hex = field(XmpHex(size=3))
-        """The LLR_INIT_ECHO sequence number. 3 bytes in hex format."""
-
-        llr_init_echo_data: Hex = field(XmpHex(size=2))
-        """The LLR_INIT_ECHO data. 2 bytes in hex format."""
-
-        llr_init_echo_mismatch: int = field(XmpLong())
-        """The number of Rx LLR_INIT_ECHO CtlOS with a sequence number or data that does not match the most recently transmitted LLR_INIT CtlOS."""
-
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get UE CtlOS Rx statistics.
-
-        :return: the UE CtlOS Rx statistics
-        :rtype: P_UE_CTLOS_RX_STATS.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-
-@register_command
-@dataclass
-class P_UE_CTLOS_TX_STATS:
-    """
-    CtlOS Tx statistics, including counters of all UE CtlOS message type values.
-    """
-
-    code: typing.ClassVar[int] = 1011
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        
-        llr_init_cnt: int = field(XmpLong())
-        """The number of LLR_INIT CtlOS transmitted."""
-
-        llr_init_echo_cnt: int = field(XmpLong())
-        """The number of LLR_INIT_ECHO CtlOS transmitted."""
-
-        llr_ack_cnt: int = field(XmpLong())
-        """The number of LLR_ACK CtlOS transmitted."""
-
-        llr_nack_cnt: int = field(XmpLong())
-        """The number of LLR_NACK CtlOS transmitted."""
-
-        llr_init_seq: Hex = field(XmpHex(size=3))
-        """The LLR_INIT sequence number. 3 bytes in hex format."""
-
-        llr_init_data: Hex = field(XmpHex(size=2))
-        """The LLR_INIT data. 2 bytes in hex format."""
-
-        llr_init_echo_seq: Hex = field(XmpHex(size=3))
-        """The LLR_INIT_ECHO sequence number. 3 bytes in hex format."""
-
-        llr_init_echo_data: Hex = field(XmpHex(size=2))
-        """The LLR_INIT_ECHO data. 2 bytes in hex format."""
-
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get UE CtlOS Tx statistics.
-
-        :return: the UE CtlOS Tx statistics
-        :rtype: P_UE_CTLOS_TX_STATS.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-
-
-@register_command
-@dataclass
-class P_UE_LINKNEG_OPTIONS:
-    """
-    This command selects which LLDP agent to use for UE Link Negotiation Options, and configures the desired LLR-W (local) setting. After the set operation, the UE Link Negotiation Options TLV will be added to the LLDPDU (if it is not already present in the LLDPDU) or replace the existing one (if Options TLV was previously added in LLDP configurataion).
-
-    The UE Link Negotiation Options TLV is used to discover and configure advanced UE link capabilities. Currently, the Options TLV is used to advertise the following UEC Link features:
-
-      * Link Layer Retry (LLR): Provides link level re-transmissions for packets with errors.
-
-    .. note::
-
-        In Link Negotiation for LLR, the R/W access of the four flags are:
-
-        * LLR-W (local): RW
-        * LLR-W (remote): RO
-        * LLR-E (local): RO
-        * LLR-E (remote): RO
-
-        When Link Negotiation is enabled on a LLDP agent, the only configurable parameter is ``LLR-W (local)``. To read ``LLR-W (remote)``, use ``P_UE_LINKNEG_OPTIONS_STATUS``. To read ``LLR-E (local)`` and ``LLR-E (remote)``, use ``P_UE_LLR_MODE``
-
-    To disable Link Negotation for UE Link Options, send the command with no parameters. This will remove the Options TLV from the LLDPDU and disable Link Negotiation for UE Link Options.
-    """
-
-    code: typing.ClassVar[int] = 1016
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        lldp_agent_index: int = field(XmpByte())
-        """integer, The LLDP agent index to use for advertising the advanced UE link capabilities. An LLDP agent must be created and configured separately using the LLDP commands before it can be used with this command. 
-        
-        Values: 0 to 7.
-        """
-
-        llr_wanted_local: UecLinkOptionLlr = field(XmpByte())
-        """coded integer, Indicates whether the local port wants LLR.
-        """
-
-    class SetDataAttr(RequestBodyStruct):
-        lldp_agent_index: int = field(XmpByte())
-        """integer, The LLDP agent index to use for advertising the advanced UE link capabilities. An LLDP agent must be created and configured separately using the LLDP commands before it can be used with this command. 
-        
-        Values: 0 to 7.
-        """
-
-        llr_wanted_local: UecLinkOptionLlr = field(XmpByte())
-        """coded integer, Indicates whether the local port wants LLR.
-        """
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the description of the port.
-
-        :return: the description of the port
-        :rtype: P_UE_LINKNEG_OPTIONS.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-    def set(self, lldp_agent_index: int, llr_wanted_local: UecLinkOptionLlr) -> Token[None]:
-        """Set the UE Link Negotiation Options.
-
-        :param lldp_agent_index: The LLDP agent index to use for advertising the advanced UE link capabilities. Values: 0 to 7.
-        :type lldp_agent_index: int
-        :param llr_wanted_local: Indicates whether the local port wants LLR. 
-        :type llr_wanted_local: UecLinkOptionLlr
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, lldp_agent_index=lldp_agent_index, llr_wanted_local=llr_wanted_local))
-
-
-@register_command
-@dataclass
-class P_UE_LINKNEG_OPTIONS_STATUS:
-    """
-    Read ``LLR-W (local)`` and ``LLR-W (remote)`` of the port, which indicates whether the local and remote ports want to use LLR. This command is for read-only purposes and does not change any configuration or behavior of the LLR.
-    """
-
-    code: typing.ClassVar[int] = 1029
-    pushed: typing.ClassVar[bool] = True
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        llr_wanted_local: UecLinkOptionLlr = field(XmpByte())
-        """coded integer, Indicates whether the local port wants LLR.
-        """
-
-        llr_wanted_remote: UecLinkOptionLlr = field(XmpByte())
-        """coded integer, Indicates whether the remote port wants LLR.
-        """
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the current UE link negotiation options status.
-
-        :return: the current UE link negotiation options status
-        :rtype: P_UE_LINKNEG_OPTIONS_STATUS.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-
-
-@register_command
-@dataclass
-class P_UE_LLR_MODE:
-    """
-    Configures the mode of operation of the LLR. Only when both local and remote ports have LLR ``ON``, the LLR will be active. If either port has LLR disabled, the LLR will be inactive.
-
-    """
-
-    code: typing.ClassVar[int] = 1013
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        llr_mode_local: OnOff = field(XmpByte())
-        """coded integer, LLR mode of the local port."""
-
-        llr_mode_remote: OnOff = field(XmpByte())
-        """coded integer, LLR mode of the remote port."""
-
-
-    class SetDataAttr(RequestBodyStruct):
-        llr_mode_local: OnOff = field(XmpByte())
-        """coded integer, LLR mode of the local port."""
-
-        llr_mode_remote: OnOff = field(XmpByte())
-        """coded integer, LLR mode of the remote port."""
-
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the LLR mode of the local port.
-
-        :return: the LLR mode of the local port
-        :rtype: P_UE_LLR_MODE.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-    def set(self, llr_mode_local: OnOff, llr_mode_remote: OnOff) -> Token[None]:
-        """Set the LLR mode of the local port.
-
-        :param llr_mode_local: the LLR mode of the local port
-        :type llr_mode_local: OnOff
-        :param llr_mode_remote: the LLR mode of the remote port
-        :type llr_mode_remote: OnOff
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, llr_mode_local=llr_mode_local, llr_mode_remote=llr_mode_remote))
-
-
-    set_on = functools.partialmethod(set, llr_mode_local=OnOff.ON, llr_mode_remote=OnOff.ON)
-    """Set LLR local to ON, LLR remote to ON."""
-
-    set_off = functools.partialmethod(set, llr_mode_local=OnOff.OFF, llr_mode_remote=OnOff.OFF)
-    """Set LLR local to OFF, LLR remote to OFF."""
-
-
-
-
-@register_command
-@dataclass
-class P_UE_LLR_RX_STATS:
-    """
-    Get the LLR Rx link-layer traffic statistics of the port. 
-    """
-
-    code: typing.ClassVar[int] = 1024
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        
-        rx_llr_eligible_good_fcs: int = field(XmpLong())
-        """The number of LLR-eligible frames received with a good FCS."""
-
-        rx_llr_eligible_poisoned_fcs: int = field(XmpLong())
-        """The number of LLR-eligible frames received with a poisoned FCS."""
-
-        rx_llr_eligible_bad_fcs: int = field(XmpLong())
-        """The number of LLR-eligible frames received with a bad FCS."""
-
-        rx_llr_eligible_total: int = field(XmpLong())
-        """The total number of LLR-eligible frames received."""
-
-        rx_llr_eligible_good_fcs_exp_seq: int = field(XmpLong())
-        """The number of LLR-eligible frames received with a good FCS that had the expected sequence number."""
-
-        rx_llr_eligible_poisoned_fcs_exp_seq: int = field(XmpLong())
-        """The number of LLR-eligible frames received with a poisoned FCS that had the expected sequence number."""
-
-        rx_llr_eligible_bad_fcs_exp_seq: int = field(XmpLong())
-        """The number of LLR-eligible frames received with a bad FCS that had the expected sequence number."""
-
-        rx_llr_eligible_total_exp_seq: int = field(XmpLong())
-        """The total number of LLR-eligible frames received with the expected sequence number (irrespective of FCS status)."""
-
-        rx_llr_eligible_missing_seq: int = field(XmpLong())
-        """The number of LLR-eligible frames received that had a sequence number that indicated a missing LLR-eligible frame in the sequence (irrespective of FCS status)."""
-
-        rx_llr_eligible_duplicate_seq: int = field(XmpLong())
-        """The number of LLR-eligible frames received that had a duplicate sequence number (irrespective of FCS status)."""
-
-        rx_llr_ineligible_good_fcs: int = field(XmpLong())
-        """The number of LLR-ineligible frames received with a good FCS."""
-
-        rx_llr_ineligible_bad_fcs: int = field(XmpLong())
-        """The number of LLR-ineligible frames received with a bad FCS."""
-
-        rx_llr_ineligible_total: int = field(XmpLong())
-        """The total number of LLR-ineligible frames received."""
-
-        rx_replay_operations: int = field(XmpLong())
-        """The number of replay operations detected."""
-
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the LLR Rx link-layer traffic statistics of the port. 
-
-        :return: the LLR Rx link-layer traffic statistics
-        :rtype: P_UE_LLR_RX_STATS.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-
-
-@register_command
-@dataclass
-class P_UE_LLR_TX_STATS:
-    """
-    Get the LLR Tx link-layer traffic statistics of the port. 
-    """
-
-    code: typing.ClassVar[int] = 1023
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        
-        tx_llr_eligible_good_fcs: int = field(XmpLong())
-        """The number of LLR-eligible frames transmitted with a good FCS."""
-
-        tx_llr_eligible_poisoned_fcs: int = field(XmpLong())
-        """The number of LLR-eligible frames transmitted with a poisoned FCS."""
-
-        tx_llr_eligible_bad_fcs: int = field(XmpLong())
-        """The number of LLR-eligible frames transmitted with a bad FCS."""
-
-        tx_llr_eligible_discard: int = field(XmpLong())
-        """The number of LLR-eligible frames discarded by the LLR TX when the TX state machine is in the INIT state and the llr_init_behavior is set to DISCARD, or when the TX state machine is in the FLUSH state and the llr_flush_behavior is set to DISCARD."""
-
-        tx_llr_eligible_total: int = field(XmpLong())
-        """The total number of LLR-eligible frames transmitted."""
-
-        tx_llr_ineligible_good_fcs: int = field(XmpLong())
-        """The number of LLR-ineligible frames transmitted with a good FCS."""
-
-        tx_llr_ineligible_bad_fcs: int = field(XmpLong())
-        """The number of LLR-ineligible frames transmitted with a bad FCS."""
-
-        tx_llr_ineligible_total: int = field(XmpLong())
-        """The total number of LLR-ineligible frames transmitted."""
-
-        tx_replayed: int = field(XmpLong())
-        """The number of frames transmitted by the LLR TX that were replayed from the LLR replay buffer."""
-
-        tx_llr_eligible_good_fcs_good_seq: int = field(XmpLong())
-        """The number of LLR-eligible frames transmitted with a good FCS and a good sequence number."""
-
-        tx_llr_eligible_poisoned_fcs_good_seq: int = field(XmpLong())
-        """The number of LLR-eligible frames transmitted with a poisoned FCS and a good sequence number."""
-
-        tx_llr_eligible_bad_fcs_good_seq: int = field(XmpLong())
-        """The number of LLR-eligible frames transmitted with a bad FCS and a good sequence number."""
-
-        tx_llr_eligible_missing_seq: int = field(XmpLong())
-        """The number of LLR-eligible frames that with missing sequence number (irrespective of the FCS)."""
-
-        tx_llr_eligible_duplicate_seq: int = field(XmpLong())
-        """The number of LLR-eligible frames that with duplicate sequence number (irrespective of the FCS)."""
-
-        tx_replay_operations: int = field(XmpLong())
-        """The number of times the LLR TX state machine has exited the REPLAY state (replay operations)."""
-
-
-    
-    def get(self) -> Token[GetDataAttr]:
-        """Get the LLR Tx link-layer traffic statistics of the port. 
-
-        :return: the LLR Tx link-layer traffic statistics
-        :rtype: P_UE_LLR_TX_STATS.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-
-
-@register_command
-@dataclass
-class P_UE_CTLOS_SPACING:
-    """
-    Configures the CtlOS spacing parameters of the port.
-    """
-
-    code: typing.ClassVar[int] = 1010
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        target_spacing: int = field(XmpInt())
-        """integer, the target CtlOS spacing."""
-
-        min_spacing: int = field(XmpInt())
-        """integer, the minimum CtlOS spacing."""
-
-        reserved1: int = field(XmpInt())
-        """integer, reserved."""
-
-        reserved2: int = field(XmpInt())
-        """integer, reserved."""
-
-    class SetDataAttr(RequestBodyStruct):
-        target_spacing: int = field(XmpInt())
-        """integer, the target CtlOS spacing."""
-
-        min_spacing: int = field(XmpInt())
-        """integer, the minimum CtlOS spacing."""
-
-        reserved1: int = field(XmpInt())
-        """integer, reserved."""
-
-        reserved2: int = field(XmpInt())
-        """integer, reserved."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the CtlOS spacing parameters of the port.
-
-        :return: the CtlOS spacing parameters of the port
-        :rtype: P_UE_CTLOS_SPACING.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-    def set(self, target_spacing: int, min_spacing: int, reserved1: int, reserved2: int) -> Token[None]:
-        """Set the CtlOS spacing parameters of the port.
-
-        :param target_spacing: the target CtlOS spacing
-        :type target_spacing: int
-        :param min_spacing: the minimum CtlOS spacing
-        :type min_spacing: int
-        :param reserved1: reserved
-        :type reserved1: int
-        :param reserved2: reserved
-        :type reserved2: int
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, target_spacing=target_spacing, min_spacing=min_spacing, reserved1=reserved1, reserved2=reserved2))
-
-
-@register_command
-@dataclass
-class P_UE_LLR_REPLAY:
-    """
-    Configures the LLR replay parameters of the port.
-    """
-
-    code: typing.ClassVar[int] = 1014
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        outstanding_seq_max: int = field(XmpInt())
-        """integer, the maximum number of outstanding sequence numbers."""
-
-        outstanding_data_max: int = field(XmpInt())
-        """integer, the maximum amount of outstanding data."""
-
-        replay_timer_max: int = field(XmpInt())
-        """integer, the maximum value of the replay timer."""
-
-        replay_ct_max: int = field(XmpInt())
-        """integer, the maximum replay count."""
-
-        pcs_lost_status_timer_max: int = field(XmpLong())
-        """long integer, the maximum value of the PCS lost status timer."""
-
-        data_age_timer_max: int = field(XmpLong())
-        """long integer, the maximum value of the data age timer."""
-
-    class SetDataAttr(RequestBodyStruct):
-        outstanding_seq_max: int = field(XmpInt())
-        """integer, the maximum number of outstanding sequence numbers."""
-
-        outstanding_data_max: int = field(XmpInt())
-        """integer, the maximum amount of outstanding data."""
-
-        replay_timer_max: int = field(XmpInt())
-        """integer, the maximum value of the replay timer."""
-
-        replay_ct_max: int = field(XmpInt())
-        """integer, the maximum replay count."""
-
-        pcs_lost_status_timer_max: int = field(XmpLong())
-        """long integer, the maximum value of the PCS lost status timer."""
-
-        data_age_timer_max: int = field(XmpLong())
-        """long integer, the maximum value of the data age timer."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the LLR replay parameters of the port.
-
-        :return: the LLR replay parameters of the port
-        :rtype: P_UE_LLR_REPLAY.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-    def set(self, outstanding_seq_max: int, outstanding_data_max: int, replay_timer_max: int, replay_ct_max: int, pcs_lost_status_timer_max: int, data_age_timer_max: int) -> Token[None]:
-        """Set the LLR replay parameters of the port.
-
-        :param outstanding_seq_max: the maximum number of outstanding sequence numbers
-        :type outstanding_seq_max: int
-        :param outstanding_data_max: the maximum amount of outstanding data
-        :type outstanding_data_max: int
-        :param replay_timer_max: the maximum value of the replay timer
-        :type replay_timer_max: int
-        :param replay_ct_max: the maximum replay count
-        :type replay_ct_max: int
-        :param pcs_lost_status_timer_max: the maximum value of the PCS lost status timer
-        :type pcs_lost_status_timer_max: int
-        :param data_age_timer_max: the maximum value of the data age timer
-        :type data_age_timer_max: int
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, outstanding_seq_max=outstanding_seq_max, outstanding_data_max=outstanding_data_max, replay_timer_max=replay_timer_max, replay_ct_max=replay_ct_max, pcs_lost_status_timer_max=pcs_lost_status_timer_max, data_age_timer_max=data_age_timer_max))
-
-
-@register_command
-@dataclass
-class P_UE_LLR_BEHAVIOR:
-    """
-    Configures the LLR behavior of the port for the INIT and FLUSH states.
-    """
-
-    code: typing.ClassVar[int] = 1015
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        llr_init_behavior: UecLlrBehavior = field(XmpByte())
-        """coded byte, the LLR behavior for LLR-eligible frames when the TX state machine is in the INIT state."""
-
-        llr_flush_behavior: UecLlrBehavior = field(XmpByte())
-        """coded byte, the LLR behavior for LLR-eligible frames when the TX state machine is in the FLUSH state."""
-
-        re_init_on_discard: YesNo = field(XmpByte())
-        """coded byte, whether to re-initialize the LLR on discard."""
-
-    class SetDataAttr(RequestBodyStruct):
-        llr_init_behavior: UecLlrBehavior = field(XmpByte())
-        """coded byte, the LLR behavior for LLR-eligible frames when the TX state machine is in the INIT state."""
-
-        llr_flush_behavior: UecLlrBehavior = field(XmpByte())
-        """coded byte, the LLR behavior for LLR-eligible frames when the TX state machine is in the FLUSH state."""
-
-        re_init_on_discard: YesNo = field(XmpByte())
-        """coded byte, whether to re-initialize the LLR on discard."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the LLR behavior of the port.
-
-        :return: the LLR behavior of the port
-        :rtype: P_UE_LLR_BEHAVIOR.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-    def set(self, llr_init_behavior: UecLlrBehavior, llr_flush_behavior: UecLlrBehavior, re_init_on_discard: YesNo) -> Token[None]:
-        """Set the LLR behavior of the port.
-
-        :param llr_init_behavior: the LLR behavior for LLR-eligible frames when the TX state machine is in the INIT state
-        :type llr_init_behavior: UecLlrBehavior
-        :param llr_flush_behavior: the LLR behavior for LLR-eligible frames when the TX state machine is in the FLUSH state
-        :type llr_flush_behavior: UecLlrBehavior
-        :param re_init_on_discard: whether to re-initialize the LLR on discard
-        :type re_init_on_discard: YesNo
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, llr_init_behavior=llr_init_behavior, llr_flush_behavior=llr_flush_behavior, re_init_on_discard=re_init_on_discard))
-
-
-@register_command
-@dataclass
-class P_UE_LLR_INIT:
-    """
-    Configures the LLR INIT parameters of the port.
-    """
-
-    code: typing.ClassVar[int] = 1017
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        init_seq: Hex = field(XmpHex(size=3))
-        """3 bytes in hex format, the 24-bit LLR_INIT sequence number."""
-
-        init_data: Hex = field(XmpHex(size=2))
-        """2 bytes in hex format, the 16-bit LLR_INIT data."""
-
-        min_spacing_multiplier: int = field(XmpInt())
-        """integer, the minimum spacing multiplier."""
-
-    class SetDataAttr(RequestBodyStruct):
-        init_seq: Hex = field(XmpHex(size=3))
-        """3 bytes in hex format, the 24-bit LLR_INIT sequence number."""
-
-        init_data: Hex = field(XmpHex(size=2))
-        """2 bytes in hex format, the 16-bit LLR_INIT data."""
-
-        min_spacing_multiplier: int = field(XmpInt())
-        """integer, the minimum spacing multiplier."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the LLR INIT parameters of the port.
-
-        :return: the LLR INIT parameters of the port
-        :rtype: P_UE_LLR_INIT.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-    def set(self, init_seq: Hex, init_data: Hex, min_spacing_multiplier: int) -> Token[None]:
-        """Set the LLR INIT parameters of the port.
-
-        :param init_seq: the 24-bit LLR_INIT sequence number (3 bytes)
-        :type init_seq: Hex
-        :param init_data: the 16-bit LLR_INIT data (2 bytes)
-        :type init_data: Hex
-        :param min_spacing_multiplier: the minimum spacing multiplier
-        :type min_spacing_multiplier: int
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, init_seq=init_seq, init_data=init_data, min_spacing_multiplier=min_spacing_multiplier))
-
-
-@register_command
-@dataclass
-class P_UE_LLR_INIT_ECHO:
-    """
-    Configures the LLR INIT_ECHO parameters of the port.
-    """
-
-    code: typing.ClassVar[int] = 1018
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        init_seq_mode: UecLlrInitEchoMode = field(XmpByte())
-        """coded byte, the source mode of the INIT_ECHO sequence number."""
-
-        init_seq: Hex = field(XmpHex(size=3))
-        """3 bytes in hex format, the 24-bit INIT_ECHO sequence number when manual mode is used."""
-
-        init_data_mode: UecLlrInitEchoMode = field(XmpByte())
-        """coded byte, the source mode of the INIT_ECHO data."""
-
-        init_data: Hex = field(XmpHex(size=2))
-        """2 bytes in hex format, the 16-bit INIT_ECHO data when manual mode is used."""
-
-    class SetDataAttr(RequestBodyStruct):
-        init_seq_mode: UecLlrInitEchoMode = field(XmpByte())
-        """coded byte, the source mode of the INIT_ECHO sequence number."""
-
-        init_seq: Hex = field(XmpHex(size=3))
-        """3 bytes in hex format, the 24-bit INIT_ECHO sequence number when manual mode is used."""
-
-        init_data_mode: UecLlrInitEchoMode = field(XmpByte())
-        """coded byte, the source mode of the INIT_ECHO data."""
-
-        init_data: Hex = field(XmpHex(size=2))
-        """2 bytes in hex format, the 16-bit INIT_ECHO data when manual mode is used."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the LLR INIT_ECHO parameters of the port.
-
-        :return: the LLR INIT_ECHO parameters of the port
-        :rtype: P_UE_LLR_INIT_ECHO.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-    def set(self, init_seq_mode: UecLlrInitEchoMode, init_seq: Hex, init_data_mode: UecLlrInitEchoMode, init_data: Hex) -> Token[None]:
-        """Set the LLR INIT_ECHO parameters of the port.
-
-        :param init_seq_mode: the source mode of the INIT_ECHO sequence number
-        :type init_seq_mode: UecLlrInitEchoMode
-        :param init_seq: the 24-bit INIT_ECHO sequence number when manual mode is used (3 bytes)
-        :type init_seq: Hex
-        :param init_data_mode: the source mode of the INIT_ECHO data
-        :type init_data_mode: UecLlrInitEchoMode
-        :param init_data: the 16-bit INIT_ECHO data when manual mode is used (2 bytes)
-        :type init_data: Hex
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, init_seq_mode=init_seq_mode, init_seq=init_seq, init_data_mode=init_data_mode, init_data=init_data))
-
-
-@register_command
-@dataclass
-class P_UE_LLR_ACKNACK:
-    """
-    Configures the LLR ACK/NACK spacing parameters of the port.
-    """
-
-    code: typing.ClassVar[int] = 1019
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        target_spacing: int = field(XmpInt())
-        """integer, the target ACK/NACK spacing."""
-
-        min_spacing: int = field(XmpInt())
-        """integer, the minimum ACK/NACK spacing."""
-
-        max_spacing: int = field(XmpInt())
-        """integer, the maximum ACK/NACK spacing."""
-
-    class SetDataAttr(RequestBodyStruct):
-        target_spacing: int = field(XmpInt())
-        """integer, the target ACK/NACK spacing."""
-
-        min_spacing: int = field(XmpInt())
-        """integer, the minimum ACK/NACK spacing."""
-
-        max_spacing: int = field(XmpInt())
-        """integer, the maximum ACK/NACK spacing."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the LLR ACK/NACK spacing parameters of the port.
-
-        :return: the LLR ACK/NACK spacing parameters of the port
-        :rtype: P_UE_LLR_ACKNACK.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-    def set(self, target_spacing: int, min_spacing: int, max_spacing: int) -> Token[None]:
-        """Set the LLR ACK/NACK spacing parameters of the port.
-
-        :param target_spacing: the target ACK/NACK spacing
-        :type target_spacing: int
-        :param min_spacing: the minimum ACK/NACK spacing
-        :type min_spacing: int
-        :param max_spacing: the maximum ACK/NACK spacing
-        :type max_spacing: int
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, target_spacing=target_spacing, min_spacing=min_spacing, max_spacing=max_spacing))
-
-
-@register_command
-@dataclass
-class P_UE_LLR_TXERR:
-    """
-    Configures the LLR TX error injection of the port.
-    """
-
-    code: typing.ClassVar[int] = 1025
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        error_type: UecLlrTxErrType = field(XmpByte())
-        """coded byte, the type of error to inject."""
-
-        pattern: UecLlrTxErrPattern = field(XmpByte())
-        """coded byte, the pattern of error injection."""
-
-        burst_size: int = field(XmpInt())
-        """integer, the number of errors in a burst."""
-
-        burst_interval: int = field(XmpInt())
-        """integer, the interval between bursts."""
-
-    class SetDataAttr(RequestBodyStruct):
-        error_type: UecLlrTxErrType = field(XmpByte())
-        """coded byte, the type of error to inject."""
-
-        pattern: UecLlrTxErrPattern = field(XmpByte())
-        """coded byte, the pattern of error injection."""
-
-        burst_size: int = field(XmpInt())
-        """integer, the number of errors in a burst."""
-
-        burst_interval: int = field(XmpInt())
-        """integer, the interval between bursts."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the LLR TX error injection configuration of the port.
-
-        :return: the LLR TX error injection configuration of the port
-        :rtype: P_UE_LLR_TXERR.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-    def set(self, error_type: UecLlrTxErrType, pattern: UecLlrTxErrPattern, burst_size: int, burst_interval: int) -> Token[None]:
-        """Set the LLR TX error injection configuration of the port.
-
-        :param error_type: the type of error to inject
-        :type error_type: UecLlrTxErrType
-        :param pattern: the pattern of error injection
-        :type pattern: UecLlrTxErrPattern
-        :param burst_size: the number of errors in a burst
-        :type burst_size: int
-        :param burst_interval: the interval between bursts
-        :type burst_interval: int
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, error_type=error_type, pattern=pattern, burst_size=burst_size, burst_interval=burst_interval))
-
-
-@register_command
-@dataclass
-class P_UE_LLR_TXERR_POISONFCS:
-    """
-    Configures the poison FCS pattern used by the LLR TX error injection of the port.
-    """
-
-    code: typing.ClassVar[int] = 1028
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        pattern: Hex = field(XmpHex(size=4))
-        """4 bytes in hex format, the 32-bit poison FCS pattern."""
-
-    class SetDataAttr(RequestBodyStruct):
-        pattern: Hex = field(XmpHex(size=4))
-        """4 bytes in hex format, the 32-bit poison FCS pattern."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the poison FCS pattern of the port.
-
-        :return: the poison FCS pattern of the port
-        :rtype: P_UE_LLR_TXERR_POISONFCS.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-    def set(self, pattern: Hex) -> Token[None]:
-        """Set the poison FCS pattern of the port.
-
-        :param pattern: the 32-bit poison FCS pattern (4 bytes)
-        :type pattern: Hex
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, pattern=pattern))
-
-
-@register_command
-@dataclass
-class P_UE_LLR_INIT_ECHO_CHK:
-    """
-    Configures the validation of the LLR INIT_ECHO sequence number and data of the port.
-    """
-
-    code: typing.ClassVar[int] = 1031
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        validate_init_seq: UecLlrEchoCheck = field(XmpByte())
-        """coded byte, whether to validate the INIT_ECHO sequence number."""
-
-        validate_init_data: UecLlrEchoCheck = field(XmpByte())
-        """coded byte, whether to validate the INIT_ECHO data."""
-
-    class SetDataAttr(RequestBodyStruct):
-        validate_init_seq: UecLlrEchoCheck = field(XmpByte())
-        """coded byte, whether to validate the INIT_ECHO sequence number."""
-
-        validate_init_data: UecLlrEchoCheck = field(XmpByte())
-        """coded byte, whether to validate the INIT_ECHO data."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the LLR INIT_ECHO validation configuration of the port.
-
-        :return: the LLR INIT_ECHO validation configuration of the port
-        :rtype: P_UE_LLR_INIT_ECHO_CHK.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-    def set(self, validate_init_seq: UecLlrEchoCheck, validate_init_data: UecLlrEchoCheck) -> Token[None]:
-        """Set the LLR INIT_ECHO validation configuration of the port.
-
-        :param validate_init_seq: whether to validate the INIT_ECHO sequence number
-        :type validate_init_seq: UecLlrEchoCheck
-        :param validate_init_data: whether to validate the INIT_ECHO data
-        :type validate_init_data: UecLlrEchoCheck
-        """
-
-        return Token(self._connection, build_set_request(self, module=self._module, port=self._port, validate_init_seq=validate_init_seq, validate_init_data=validate_init_data))
-
-
-@register_command
-@dataclass
-class P_UE_LLR_TXFSM_STATE:
-    """
-    Return the current state of the LLR Tx FSM that the port is in, along with the current values of the relevant timers and counters for that state. The possible states are: LLR_OFF, INIT, ADVANCE, REPLAY, and FLUSH. The specific timers and counters returned depend on the current state of the FSM.
-    """
-
-    code: typing.ClassVar[int] = 1021
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        txfsm_state: UecLlrTxFsmState = field(XmpByte())
-        """coded byte, the current state of the LLR Tx FSM that the port is in."""
-
-        ns_in_llr_off: int = field(XmpLong())
-        """long integer, the number of nanoseconds the FSM has been in the LLR_OFF state since the last time it was cleared."""
-
-        ns_in_init: int = field(XmpLong())
-        """long integer, the number of nanoseconds the FSM has been in the INIT state since the last time it was cleared."""
-
-        ns_in_advance: int = field(XmpLong())
-        """long integer, the number of nanoseconds the FSM has been in the ADVANCE state since the last time it was cleared."""
-
-        ns_in_replay: int = field(XmpLong())
-        """long integer, the number of nanoseconds the FSM has been in the REPLAY state since the last time it was cleared."""
-
-        ns_in_flush: int = field(XmpLong())
-        """long integer, the number of nanoseconds the FSM has been in the FLUSH state since the last time it was cleared."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the state of the LLR TX finite state machine of the port.
-
-        :return: the state of the LLR TX finite state machine of the port
-        :rtype: P_UE_LLR_TXFSM_STATE.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-
-@register_command
-@dataclass
-class P_UE_LLR_RXFSM_STATE:
-    """
-    Return the current state of the LLR Rx FSM that the port is in, along with the current values of the relevant timers and counters for that state. The possible states are: OFF, SEND_ACKS, SEND_NACK, and NACK_SENT. The specific timers and counters returned depend on the current state of the FSM.
-    """
-
-    code: typing.ClassVar[int] = 1022
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        rxfsm_state: UecLlrRxFsmState = field(XmpByte())
-        """coded byte, the current state of the LLR Rx FSM that the port is in."""
-
-        ns_in_off: int = field(XmpLong())
-        """long integer, the number of nanoseconds the FSM has been in the OFF state since the last time it was cleared."""
-
-        ns_in_send_acks: int = field(XmpLong())
-        """long integer, the number of nanoseconds the FSM has been in the SEND_ACKS state since the last time it was cleared."""
-
-        ns_in_send_nack: int = field(XmpLong())
-        """long integer, the number of nanoseconds the FSM has been in the SEND_NACK state since the last time it was cleared."""
-
-        ns_in_nack_sent: int = field(XmpLong())
-        """long integer, the number of nanoseconds the FSM has been in the NACK_SENT state since the last time it was cleared."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the state of the LLR RX finite state machine of the port.
-
-        :return: the state of the LLR RX finite state machine of the port
-        :rtype: P_UE_LLR_RXFSM_STATE.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-
-@register_command
-@dataclass
-class P_UE_LLR_STATUS:
-    """
-    Get the active status of the LLR of the port.
-    """
-
-    code: typing.ClassVar[int] = 1030
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        status: YesNo = field(XmpByte())
-        """coded byte, whether the LLR is active."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the active status of the LLR of the port.
-
-        :return: the active status of the LLR of the port
-        :rtype: P_UE_LLR_STATUS.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-
-@register_command
-@dataclass
-class P_UE_CTLOS_TX_INTERVAL:
-    """
-    Get the CtlOS Tx interval statistics of the port. For each CtlOS message type, the minimum, maximum, and average intervals are reported.
-    """
-
-    code: typing.ClassVar[int] = 1032
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        ctlos_min: int = field(XmpLong())
-        """long integer, the minimum CtlOS Tx interval."""
-
-        ctlos_max: int = field(XmpLong())
-        """long integer, the maximum CtlOS Tx interval."""
-
-        ctlos_avg: int = field(XmpLong())
-        """long integer, the average CtlOS Tx interval."""
-
-        llr_init_min: int = field(XmpLong())
-        """long integer, the minimum LLR_INIT Tx interval."""
-
-        llr_init_max: int = field(XmpLong())
-        """long integer, the maximum LLR_INIT Tx interval."""
-
-        llr_init_avg: int = field(XmpLong())
-        """long integer, the average LLR_INIT Tx interval."""
-
-        llr_init_echo_min: int = field(XmpLong())
-        """long integer, the minimum LLR_INIT_ECHO Tx interval."""
-
-        llr_init_echo_max: int = field(XmpLong())
-        """long integer, the maximum LLR_INIT_ECHO Tx interval."""
-
-        llr_init_echo_avg: int = field(XmpLong())
-        """long integer, the average LLR_INIT_ECHO Tx interval."""
-
-        llr_ack_min: int = field(XmpLong())
-        """long integer, the minimum LLR_ACK Tx interval."""
-
-        llr_ack_max: int = field(XmpLong())
-        """long integer, the maximum LLR_ACK Tx interval."""
-
-        llr_ack_avg: int = field(XmpLong())
-        """long integer, the average LLR_ACK Tx interval."""
-
-        llr_nack_min: int = field(XmpLong())
-        """long integer, the minimum LLR_NACK Tx interval."""
-
-        llr_nack_max: int = field(XmpLong())
-        """long integer, the maximum LLR_NACK Tx interval."""
-
-        llr_nack_avg: int = field(XmpLong())
-        """long integer, the average LLR_NACK Tx interval."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the CtlOS Tx interval statistics of the port.
-
-        :return: the CtlOS Tx interval statistics of the port
-        :rtype: P_UE_CTLOS_TX_INTERVAL.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-
-@register_command
-@dataclass
-class P_UE_CTLOS_RX_INTERVAL:
-    """
-    Get the CtlOS Rx interval statistics of the port. For each CtlOS message type, the minimum, maximum, and average intervals are reported.
-    """
-
-    code: typing.ClassVar[int] = 1033
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        ctlos_min: int = field(XmpLong())
-        """long integer, the minimum CtlOS Rx interval."""
-
-        ctlos_max: int = field(XmpLong())
-        """long integer, the maximum CtlOS Rx interval."""
-
-        ctlos_avg: int = field(XmpLong())
-        """long integer, the average CtlOS Rx interval."""
-
-        llr_init_min: int = field(XmpLong())
-        """long integer, the minimum LLR_INIT Rx interval."""
-
-        llr_init_max: int = field(XmpLong())
-        """long integer, the maximum LLR_INIT Rx interval."""
-
-        llr_init_avg: int = field(XmpLong())
-        """long integer, the average LLR_INIT Rx interval."""
-
-        llr_init_echo_min: int = field(XmpLong())
-        """long integer, the minimum LLR_INIT_ECHO Rx interval."""
-
-        llr_init_echo_max: int = field(XmpLong())
-        """long integer, the maximum LLR_INIT_ECHO Rx interval."""
-
-        llr_init_echo_avg: int = field(XmpLong())
-        """long integer, the average LLR_INIT_ECHO Rx interval."""
-
-        llr_ack_min: int = field(XmpLong())
-        """long integer, the minimum LLR_ACK Rx interval."""
-
-        llr_ack_max: int = field(XmpLong())
-        """long integer, the maximum LLR_ACK Rx interval."""
-
-        llr_ack_avg: int = field(XmpLong())
-        """long integer, the average LLR_ACK Rx interval."""
-
-        llr_nack_min: int = field(XmpLong())
-        """long integer, the minimum LLR_NACK Rx interval."""
-
-        llr_nack_max: int = field(XmpLong())
-        """long integer, the maximum LLR_NACK Rx interval."""
-
-        llr_nack_avg: int = field(XmpLong())
-        """long integer, the average LLR_NACK Rx interval."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the CtlOS Rx interval statistics of the port.
-
-        :return: the CtlOS Rx interval statistics of the port
-        :rtype: P_UE_CTLOS_RX_INTERVAL.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
-
-
-@register_command
-@dataclass
-class P_UE_CTLOS_RX_ERRORS:
-    """
-    Get the CtlOS Rx error counters of the port.
-    """
-
-    code: typing.ClassVar[int] = 1034
-    pushed: typing.ClassVar[bool] = False
-
-    _connection: 'interfaces.IConnection'
-    _module: int
-    _port: int
-
-    class GetDataAttr(ResponseBodyStruct):
-        ctlos_msg_type_errors: int = field(XmpLong())
-        """long integer, the number of CtlOS message type errors."""
-
-        ctlos_padding_errors: int = field(XmpLong())
-        """long integer, the number of CtlOS padding errors."""
-
-        llr_init_echo_invalid_init_seq: int = field(XmpLong())
-        """long integer, the number of LLR_INIT_ECHO with an invalid INIT sequence number."""
-
-        llr_init_echo_invalid_init_data: int = field(XmpLong())
-        """long integer, the number of LLR_INIT_ECHO with invalid INIT data."""
-
-        llr_init_interval_low: int = field(XmpLong())
-        """long integer, the number of LLR_INIT received with an interval that is too low."""
-
-        llr_ack_nack_interval_high: int = field(XmpLong())
-        """long integer, the number of LLR_ACK/NACK received with an interval that is too high."""
-
-        llr_ack_nack_interval_low: int = field(XmpLong())
-        """long integer, the number of LLR_ACK/NACK received with an interval that is too low."""
-
-        llr_ctlos_padding_error: int = field(XmpLong())
-        """long integer, the number of LLR CtlOS padding errors."""
-
-    def get(self) -> Token[GetDataAttr]:
-        """Get the CtlOS Rx error counters of the port.
-
-        :return: the CtlOS Rx error counters of the port
-        :rtype: P_UE_CTLOS_RX_ERRORS.GetDataAttr
-        """
-
-        return Token(self._connection, build_get_request(self, module=self._module, port=self._port))
 
 
 
@@ -8265,76 +4896,5 @@ __all__ = [
     "P_XMITONE",
     "P_XMITONETIME",
     "P_FAULTCNT",
-    "P_MACSEC_TXSC_CREATE",
-    "P_MACSEC_TXSC_INDICES",
-    "P_MACSEC_TXSC_DELETE",
-    "P_MACSEC_TXSC_CONF_OFFSET",
-    "P_MACSEC_TXSC_DESCR",
-    "P_MACSEC_TXSC_SCI_MODE",
-    "P_MACSEC_TXSC_SCI",
-    "P_MACSEC_TXSC_CIPHERSUITE",
-    "P_MACSEC_TXSC_STARTING_PN",
-    "P_MACSEC_TXSC_REKEY_MODE",
-    "P_MACSEC_TXSC_ENCRYPT",
-    "P_MACSEC_TXSC_SAK_VALUE",
-    "P_MACSEC_TXSC_XPN_SSCI",
-    "P_MACSEC_TXSC_XPN_SALT",
-    "P_MACSEC_RXSC_CREATE",
-    "P_MACSEC_RXSC_INDICES",
-    "P_MACSEC_RXSC_DELETE",
-    "P_MACSEC_RXSC_DESCR",
-    "P_MACSEC_RXSC_SCI",
-    "P_MACSEC_RXSC_CONF_OFFSET",
-    "P_MACSEC_RXSC_CIPHERSUITE",
-    "P_MACSEC_RXSC_LOWEST_PN",
-    "P_MACSEC_RXSC_TPLDID",
-    "P_MACSEC_RXSC_SAK_VALUE",
-    "P_MACSEC_RXSC_XPN_SSCI",
-    "P_MACSEC_RXSC_XPN_SALT",
-    "P_MACSEC_TX_STATS",
-    "P_MACSEC_TXSC_STATS",
-    "P_MACSEC_TX_CLEAR",
-    "P_MACSEC_RX_STATS",
-    "P_MACSEC_RXSC_STATS",
-    "P_MACSEC_RX_CLEAR",
-    "P_MACSEC_RX_ENABLE",
-    "P_MACSEC_RXSC_AN",
-    "P_MACSEC_RXSC_NEXT_PN",
-    "P_MACSEC_TXSC_NEXT_PN",
-    "P_MACSEC_RXSC_PN",
-    "P_MACSEC_TXSC_NEXT_AN",
-    "P_LLDP_CLEAR",
-    "P_LLDP_CONFIG",
-    "P_LLDP_CREATE",
-    "P_LLDP_DATA",
-    "P_LLDP_DELETE",
-    "P_LLDP_HEADER",
-    "P_LLDP_INDICES",
-    "P_LLDP_NEIGHBORS",
-    "P_LLDP_OPMODE",
-    "P_LLDP_STATS",
     "P_EDUN_RX_STATUS",
-    "P_UE_CTLOS_CLEAR",
-    "P_UE_CTLOS_RX_STATS",
-    "P_UE_CTLOS_TX_STATS",
-    "P_UE_LINKNEG_OPTIONS",
-    "P_UE_LINKNEG_OPTIONS_STATUS",
-    "P_UE_LLR_MODE",
-    "P_UE_LLR_RX_STATS",
-    "P_UE_LLR_TX_STATS",
-    "P_UE_CTLOS_SPACING",
-    "P_UE_LLR_REPLAY",
-    "P_UE_LLR_BEHAVIOR",
-    "P_UE_LLR_INIT",
-    "P_UE_LLR_INIT_ECHO",
-    "P_UE_LLR_ACKNACK",
-    "P_UE_LLR_TXERR",
-    "P_UE_LLR_TXERR_POISONFCS",
-    "P_UE_LLR_INIT_ECHO_CHK",
-    "P_UE_LLR_TXFSM_STATE",
-    "P_UE_LLR_RXFSM_STATE",
-    "P_UE_LLR_STATUS",
-    "P_UE_CTLOS_TX_INTERVAL",
-    "P_UE_CTLOS_RX_INTERVAL",
-    "P_UE_CTLOS_RX_ERRORS",
 ]
