+++ xoa_driver/internals/commands/pc_commands.py
@@ -1,4 +1,4 @@
-"""Port Capture Commands"""
+"""Port Commands - PCAP Capture"""
 from __future__ import annotations
 from dataclasses import dataclass
 import typing
