+++ xoa_driver/internals/commands/ped_commands.py
@@ -1,4 +1,4 @@
-"""Port (Impairment) Distribution Commands"""
+"""Port Commands - Impairment Distribution"""
 from __future__ import annotations
 from dataclasses import dataclass
 import typing
