+++ xoa_driver/internals/commands/pf_commands.py
@@ -1,4 +1,4 @@
-"""Port Filter Commands"""
+"""Port Commands - Filter"""
 from __future__ import annotations
 from dataclasses import dataclass
 import typing
