+++ xoa_driver/internals/commands/pl1_commands.py
@@ -1,4 +1,4 @@
-"""Port Layer 1 Commands"""
+"""Port Commands - Layer 1"""
 from __future__ import annotations
 from dataclasses import dataclass
 import typing
