+++ xoa_driver/internals/commands/pl_commands.py
@@ -1,4 +1,4 @@
-"""Port Length Term Commands"""
+"""Port Commands - Length Term"""
 from __future__ import annotations
 from dataclasses import dataclass
 import typing
