+++ xoa_driver/internals/commands/pm_commands.py
@@ -1,4 +1,4 @@
-"""Port Match Term Commands"""
+"""Port Commands - Match Term"""
 from __future__ import annotations
 from dataclasses import dataclass
 import typing
