+++ xoa_driver/internals/commands/pp_commands.py
@@ -1,4 +1,4 @@
-"""Port Basic Layer 1 Commands"""
+"""Port Commands - Basic Layer 1"""
 from __future__ import annotations
 from dataclasses import dataclass
 import typing
